### 2023-12-03 #6112 @editorialbot
[PRE REVIEW]: TelescopeML -- I. An End-to-End Python Package for Interpreting Telescope Datasets through Training  Machine Learning Models, Generating Statistical Reports, and Visualizing Results
**Submitting author:** @EhsanGharibNezhad (<a href="http://orcid.org/0000-0002-4088-7262">Ehsan Gharib-Nezhad</a>)
**Repository:** https://github.com/EhsanGharibNezhad/TelescopeML
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.3
