### 2024-01-04 e04e221
update DeepTrainer.py: CNN Architechiture, output layer, units=1
