+++ TelescopeML/DeepTrainer.py
@@ -37,6 +37,296 @@ from tensorflow import keras
 from tensorflow.keras.models import save_model
 
 
+
+class TrainRegressorCNN_v2:
+    """
+    Train Convolutional Neural Networks model using regression approach
+
+    Parameters
+    -----------
+    X1_train : array
+        Row-StandardScaled input spectra for training.
+    X1_val : array
+        Row-StandardScaled input spectra for validation.
+    X1_test : array
+        Row-StandardScaled input spectra for testing.
+
+    X2_train : array
+        Col-StandardScaled Mix Max of all rows of input spectra for training.
+    X2_val : array
+        Col-StandardScaled Mix Max of all rows of input spectra for validation.
+    X2_test : array
+        Col-StandardScaled Mix Max of all rows of input spectra for testing.
+
+    y1_train : array
+        Col-StandardScaled target feature 1 for training.
+    y1_val : array
+        Col-StandardScaled target feature 1 for validation.
+    y1_test : array
+        Col-StandardScaled target feature 1 for testing.
+
+    y2_train : array
+        Col-StandardScaled target feature 2 for training.
+    y2_val : array
+        Col-StandardScaled target feature 2 for validation.
+    y2_test : array
+        Col-StandardScaled target feature 2 for testing.
+
+    y3_train : array
+        Col-StandardScaled target feature 3 for training.
+    y3_val : array
+        Col-StandardScaled target feature 3 for validation.
+    y3_test : array
+        Col-StandardScaled target feature 3 for testing.
+
+    y4_train : array
+        Col-StandardScaled target feature 4 for training.
+    y4_val : array
+        Col-StandardScaled target feature 4 for validation.
+    y4_test : array
+        Col-StandardScaled target feature 4 for testing.
+    """
+    def __init__(self,
+                 X1_train, X1_val, X1_test,  # Row-StandardScaled input spectra
+                 X2_train, X2_val, X2_test,  # Col-StandardScaled Mix Max of all rows of input spetra
+                 y1_train, y1_val, y1_test,  # Col-StandardScaled target feature 1
+                 y2_train, y2_val, y2_test,  # Col-StandardScaled target feature 2
+                 y3_train, y3_val, y3_test,  # Col-StandardScaled target feature 3
+                 y4_train, y4_val, y4_test,  # Col-StandardScaled target feature 4
+                 ):
+
+        # train, val, test sets for input 1 (main 104 spectral features)
+        self.X1_train, self.X1_val, self.X1_test = X1_train, X1_val, X1_test
+
+        # train, val, test sets for input 2 (Min and Max 2 features)
+        self.X2_train, self.X2_val, self.X2_test = X2_train, X2_val, X2_test
+
+        # train, val, test sets for target features
+        self.y1_train, self.y1_val, self.y1_test = y1_train, y1_val, y1_test
+        self.y2_train, self.y2_val, self.y2_test = y2_train, y2_val, y2_test
+        self.y3_train, self.y3_val, self.y3_test = y3_train, y3_val, y3_test
+        self.y4_train, self.y4_val, self.y4_test = y4_train, y4_val, y4_test
+
+    def build_model(self,
+                    config, # dic
+                    ):
+        """
+        Build a CNN model with the given hyperparameters.
+
+        Parameters
+        ----------
+        hyperparameters : dict
+            A dictionary containing hyperparameter settings.
+
+            hyperparameters keys includes:
+                - 'Conv__num_blocks' (int): Number of blocks in the CNN model.
+                - 'Conv__num_layers_per_block' (int): Number of layers in each convolutional block.
+                - 'Conv__num_filters' (list): Number of filters in each layer.
+                - 'Conv__kernel_size' (int): Size of the convolutional kernel.
+                - 'Conv__MaxPooling1D' (bool): MaxPooling1D size.
+
[…]
