+++ TelescopeML/DeepTuner.py
@@ -1,0 +1,609 @@
+# Import functions from other modules ============================
+# from io_funs import LoadSave
+
+# Import python libraries ========================================
+
+# ******* Standard Data Manipulation / Statistical Libraries *****
+import pandas as pd
+pd.options.mode.chained_assignment = None  # Suppress warnings
+import numpy as np
+import pickle as pk
+
+from typing import Union
+
+
+# ******* Data Visualization Libraries ****************************
+
+import matplotlib.pyplot as plt
+
+import seaborn as sns
+
+from bokeh.io import output_notebook
+output_notebook()
+from bokeh.plotting import show,figure
+TOOLTIPS = [
+    ("index", "$index"),
+    ("(x,y)", "($x, $y)"),
+]
+
+# ******** Data science / Machine learning Libraries ***************
+import tensorflow as tf
+from tensorflow.keras.layers import Conv1D, MaxPooling1D, Flatten, Dense, Dropout
+from tensorflow.keras.callbacks import EarlyStopping
+
+tf.get_logger().setLevel('ERROR')
+
+from tensorflow import keras
+from tensorflow.keras.models import save_model
+
+
+
+class TrainRegressorCNN_v2:
+    """
+    Train Convolutional Neural Networks model using regression approach
+
+    Parameters
+    -----------
+    X1_train : array
+        Row-StandardScaled input spectra for training.
+    X1_val : array
+        Row-StandardScaled input spectra for validation.
+    X1_test : array
+        Row-StandardScaled input spectra for testing.
+
+    X2_train : array
+        Col-StandardScaled Mix Max of all rows of input spectra for training.
+    X2_val : array
+        Col-StandardScaled Mix Max of all rows of input spectra for validation.
+    X2_test : array
+        Col-StandardScaled Mix Max of all rows of input spectra for testing.
+
+    y1_train : array
+        Col-StandardScaled target feature 1 for training.
+    y1_val : array
+        Col-StandardScaled target feature 1 for validation.
+    y1_test : array
+        Col-StandardScaled target feature 1 for testing.
+
+    y2_train : array
+        Col-StandardScaled target feature 2 for training.
+    y2_val : array
+        Col-StandardScaled target feature 2 for validation.
+    y2_test : array
+        Col-StandardScaled target feature 2 for testing.
+
+    y3_train : array
+        Col-StandardScaled target feature 3 for training.
+    y3_val : array
+        Col-StandardScaled target feature 3 for validation.
+    y3_test : array
+        Col-StandardScaled target feature 3 for testing.
+
+    y4_train : array
+        Col-StandardScaled target feature 4 for training.
+    y4_val : array
+        Col-StandardScaled target feature 4 for validation.
+    y4_test : array
+        Col-StandardScaled target feature 4 for testing.
+    """
+    def __init__(self,
+                 X1_train, X1_val, X1_test,  # Row-StandardScaled input spectra
+                 X2_train, X2_val, X2_test,  # Col-StandardScaled Mix Max of all rows of input spetra
+                 y1_train, y1_val, y1_test,  # Col-StandardScaled target feature 1
+                 y2_train, y2_val, y2_test,  # Col-StandardScaled target feature 2
+                 y3_train, y3_val, y3_test,  # Col-StandardScaled target feature 3
+                 y4_train, y4_val, y4_test,  # Col-StandardScaled target feature 4
+                 ):
+
+        # train, val, test sets for input 1 (main 104 spectral features)
+        self.X1_train, self.X1_val, self.X1_test = X1_train, X1_val, X1_test
+
+        # train, val, test sets for input 2 (Min and Max 2 features)
[…]
