+++ TelescopeML/DeepTuner.py
@@ -1,91 +1,104 @@
+import scipy.signal.signaltools
+
+
+def _centered(arr, newsize):
+    # Return the center newsize portion of the array.
+    newsize = np.asarray(newsize)
+    currsize = np.array(arr.shape)
+    startind = (currsize - newsize) // 2
+    endind = startind + newsize
+    myslice = [slice(startind[k], endind[k]) for k in range(len(endind))]
+    return arr[tuple(myslice)]
+
+
+scipy.signal.signaltools._centered = _centered
+
+import os
+
+# os.environ['NUMEXPR_MAX_THREADS'] = '42'
+# os.environ['NUMEXPR_NUM_THREADS'] = '40'
+# import numexpr as ne
+
+ip_address = '127.0.0.6'
+
+from TelescopeML.DataMaster import *
+from TelescopeML.Predictor import *
+from TelescopeML.StatVisAnalyzer import *
+from TelescopeML.IO_utils import load_or_dump_trained_model_CNN
+
+import logging
+
+logging.basicConfig(level=logging.DEBUG)
+
+# Libraries for BOHB Package
+import argparse
+
+import hpbandster.core.nameserver as hpns
+import hpbandster.core.result as hpres
+
+from hpbandster.optimizers import BOHB as BOHB
+# from hpbandster.examples.commons import MyWorker
+
+# import logging
+# logging.basicConfig(level=logging.ERROR) # Change the level to INFO or higher
+# import argparse
+
+# import hpbandster.core.nameserver as hpns
+# import hpbandster.core.result as hpres
+
+# from hpbandster.optimizers import BOHB as BOHB
+# from hpbandster.examples.commons import MyWorker
+
+import numpy
+import time
+
+import ConfigSpace as CS
+from hpbandster.core.worker import Worker
+
 # Import functions from other modules ============================
 # from io_funs import LoadSave
 
 # Import python libraries ========================================
-
-# ******* Standard Data Manipulation / Statistical Libraries *****
-import pandas as pd
-pd.options.mode.chained_assignment = None  # Suppress warnings
 import numpy as np
-import pickle as pk
-
-from typing import Union
-
-
-# ******* Data Visualization Libraries ****************************
-
 import matplotlib.pyplot as plt
 
-import seaborn as sns
+# Import BOHB libraries ========================================
 
-from bokeh.io import output_notebook
-output_notebook()
-from bokeh.plotting import show,figure
-TOOLTIPS = [
-    ("index", "$index"),
-    ("(x,y)", "($x, $y)"),
-]
 
-# ******** Data science / Machine learning Libraries ***************
+import ConfigSpace.hyperparameters as CSH
+
+# import logging
+# logging.basicConfig(level=logging.DEBUG)
+
+
 import tensorflow as tf
-from tensorflow.keras.layers import Conv1D, MaxPooling1D, Flatten, Dense, Dropout
-from tensorflow.keras.callbacks import EarlyStopping
-
-tf.get_logger().setLevel('ERROR')
-
 from tensorflow import keras
-from tensorflow.keras.models import save_model
+from tensorflow.keras.models import Sequential
+from tensorflow.keras.layers import Dense, Flatten
+from tensorflow.keras.layers import Conv1D, Reshape
+from tensorflow.keras.callbacks import EarlyStopping, ReduceLROnPlateau
+from tensorflow.keras.utils import plot_model
+
+from tensorflow.keras.layers import Input, Conv1D, Concatenate, Dense, MaxPooling1D, Dropout
+from tensorflow.keras.models import Model
+
+# mlp for multi-output regression
+from numpy import mean
+from numpy import std
+from keras.models import Sequential
+from keras.layers import Dense
+
[…]
