### 2024-01-24 #6112 @plaplant
@EhsanGharibNezhad thanks for looking for potential reviewers! It definitely looks like @mdhabibi has the technical qualifications to be a good reviewer. However, my knee-jerk reaction is that this might not comply with JOSS's conflict of interest policy (a summary of which is available [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)). If @mdhabibi is a current collaborator/contributor, or works with you at the same institution or on the same project, then there is enough of a conflict of interest that the review might not be impartial. On the other hand, if @mdhabibi is merely someone you know (of) professionally and do not interact with directly, then it might not be a conflict of interest.

If you don't mind sharing a bit more about your working relationship with @mdhabibi, that would help determine if there is a conflict of interest or not. Meanwhile, I will continue looking for potential reviewers on my end, as we need 2 to proceed with the review. Thanks for your help!
