### 2024-01-26 #6112 @EhsanGharibNezhad
@plaplant Thanks for sharing these details. I was hesitant at first when I mentioned his name. He is not affiliated with the same institution where I work, but he is interested in collaborating with me on a project using the TelescopeML package by adding a module to it. Therefore, it is crucial for him to become an expert in the package and ensure that it is free of any bugs. This makes us collaborators in the near term.

To be precise, he is going to review and eventually utilize this package, whether he is assigned as a JOSS reviewer or not. He needs to ensure that the Sphinx instructions are clear, and the modules are fully functional. However, I thought assigning him as a reviewer would facilitate this process.

Anyway, this is my first submission to JOSS, and you are more familiar with the rules.
