### 2024-01-30 #6112 @EhsanGharibNezhad
Here is another astronomer/Python developer who is willing to take the time and help improve the quality of this package.
Name: Nick Wogan [@ Nicholaswogan]
Github: https://github.com/Nicholaswogan
Email: nicholas.f.wogan@nasa.gov
