### 2024-02-08 #6112 @EhsanGharibNezhad
Hi @plaplant @dfm @jedbrown, do you have any update?

We are about to submit a second paper for a proof of concept and need to have this JOSS paper at least in progress.
