### 2024-02-08 #6112 @arfon
@EhsanGharibNezhad please be patient. JOSS editors are volunteers and the editor has informed you that he is seeking reviewers for your submission.

Were this any other journal you would have zero visibility into the exact status of your paper. JOSS runs an open review process because we believe it's a better one for all, but it requires patience and grace from all involved.
