### 2024-02-09 #6112 @EhsanGharibNezhad
@arfon, my intention was to facilitate the process by recommending a few potential reviewers in the field, as per your assigned editor's suggestion earlier. Of course, I'm grateful for the time that you and other members of the JOSS editorial team devote in this process.
