### 2024-02-09 #6112 @plaplant
@EhsanGharibNezhad thank you for your suggestions for potential reviewers. Unfortunately, it looks like they also have conflicts of interest. In the paper, your affiliations include NASA Ames Research Center, and the suggested reviewers also seem to share that affiliation. [As stated in the JOSS conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), being employed by the same institution in the recent past (defined as one year) is considered a conflict of interest, with the reasoning being that it may lead to a perception that the review is not impartial.

I have been looking for reviewers on my end, but have not been updating here because no one has concretely agreed to review yet. As I confirm reviewers, I will add them to this issue. Although we typically wait for 2 confirmed reviewers before beginning the review, I am willing to start the review with just a single reviewer while working to confirm a second one, which I have done occasionally on previous submissions. I understand you are anxious to begin the review process, but part of this depends on lining up qualified reviewers (who are also volunteers). I appreciate your patience, and will provide updates as they develop.
