### 2024-02-15 #6346 @oparisot
Hello @EhsanGharibNezhad,

I will go deeply into the review, but before, you should add a short summary in the paper, as required in JOSS guidelines: "A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience."

Moreover, you should clarifiy the 'Functionality and Key Features'. The current paper proposes a high-level list of features, but it is too vague. You should describe the inputs and outputs (files formats? time series?), you should present the goal of AI models (what do you mean by 'predicting atmospheric parameters'?),you should describe the type of CNN, etc.

Finally, you should position and compare your software to existing tools (ex https://github.com/MartianColonist/POSEIDON, if it not appropriate then please list tools that are 'comparable' to your software.

I will test the software in few day.

Thank you,

Olivier
