### 2024-02-27 2557f3c
create new module, DeepTuner, to utilize BOHB and do hyperparameter tunings for CNNs
