+++ TelescopeML/DeepTuner.py
@@ -19,7 +19,7 @@ import os
 # os.environ['NUMEXPR_NUM_THREADS'] = '40'
 # import numexpr as ne
 
-ip_address = '127.0.0.6'
+ip_address = '127.0.0.1'
 
 from TelescopeML.DataMaster import *
 from TelescopeML.Predictor import *
@@ -494,19 +494,19 @@ import sys
 # Check if running as a Jupyter kernel
 if "ipykernel_launcher" in sys.argv[0]:
     args = argparse.Namespace(
-        min_budget=15,
-        max_budget=30,
-        n_iterations=50000,
+        min_budget=3,
+        max_budget=5,
+        n_iterations=5,
         n_workers=40,
         worker=False
     )
 else:
     # Create the optimization object with the desired queue size
     parser = argparse.ArgumentParser(description='Example 3 - Local and Parallel Execution.')
-    parser.add_argument('--min_budget', type=float, help='Minimum budget used during the optimization.', default=20)
-    parser.add_argument('--max_budget', type=float, help='Maximum budget used during the optimization.', default=50)
+    parser.add_argument('--min_budget', type=float, help='Minimum budget used during the optimization.', default=3)
+    parser.add_argument('--max_budget', type=float, help='Maximum budget used during the optimization.', default=5)
     parser.add_argument('--n_iterations', type=int, help='Number of iterations performed by the optimizer',
-                        default=50000)
+                        default=5)
     parser.add_argument('--n_workers', type=int, help='Number of workers to run in parallel.', default=40)
     parser.add_argument('--worker', help='Flag to turn this into a worker process', action='store_true')
 
@@ -519,6 +519,7 @@ else:
 # Note the run_id argument. This uniquely identifies a run of any HpBandSter optimizer.
 
 NS = hpns.NameServer(run_id='example1', host=ip_address, port=None)
+
 NS.start()
 
 # Step 2: Start a worker
@@ -582,13 +583,466 @@ NS.shutdown()
 # It holds informations about the optimization run like the incumbent (=best) configuration.
 # For further details about the Result object, see its documentation.
 # Here we simply print out the best config and some statistics about the performed runs.
-id2config = res.get_id2config_mapping()
-incumbent = res.get_incumbent_id()
+# id2config = res.get_id2config_mapping()
+# incumbent = res.get_incumbent_id()
+#
+# print('Best found configuration:', id2config[incumbent]['config'])
+# print('A total of %i unique configurations where sampled.' % len(id2config.keys()))
+# print('A total of %i runs where executed.' % len(res.get_all_runs()))
+# print('Total budget corresponds to %.1f full function evaluations.' % (
+#             sum([r.budget for r in res.get_all_runs()]) / args.max_budget))
+# print('The run took  %.1f seconds to complete.' % (
+#             all_runs[-1].time_stamps['finished'] - all_runs[0].time_stamps['started']))
 
-print('Best found configuration:', id2config[incumbent]['config'])
-print('A total of %i unique configurations where sampled.' % len(id2config.keys()))
-print('A total of %i runs where executed.' % len(res.get_all_runs()))
-print('Total budget corresponds to %.1f full function evaluations.' % (
-            sum([r.budget for r in res.get_all_runs()]) / args.max_budget))
-print('The run took  %.1f seconds to complete.' % (
-            all_runs[-1].time_stamps['finished'] - all_runs[0].time_stamps['started']))
+
+import pandas as pd
+from bokeh.plotting import figure, show
+from bokeh.io import output_notebook
+from bokeh.models import LogTicker, LogTickFormatter
+
+
+class Check_bohb_results:
+
+    def __init__(self, path):
+        self.path = path
+
+    def read_bohb_results(self, __print_results__=True):
+
[…]
