### 2024-02-27 d05f8be
update DeepTrainer.py over the writing the ApJ paper on Jan-Feb
