### 2024-02-27 ad970d2
update Predictor.py over the writing the ApJ paper on Jan-Feb
