+++ TelescopeML/StatVisAnalyzer.py
@@ -21,12 +21,22 @@ import pprint
 # ******* Data Visulaization Libraries ****************************
 import seaborn as sns
 import matplotlib.pyplot as plt
+from matplotlib.ticker import (MultipleLocator, AutoMinorLocator)
 
 
 from bokeh.plotting import output_notebook
 
 from bokeh.plotting import figure, show
 from bokeh.models import ColumnDataSource
+from bokeh.io import export_svg
+
+import json
+import pandas as pd
+
+import warnings
+
+# Filter out the specific FutureWarning
+warnings.filterwarnings("ignore", category=FutureWarning, module="seaborn._oldcore")
 
 
 def print_results_fun(targets, print_title=None):
@@ -146,7 +156,7 @@ def regression_report(trained_model,
                 r2_score_train, rmse_score_test, mean_train, std_train))
 
         # Plot histograms of residuals
-        axs[0].set_title(['Gravity', 'C_O_ratio', 'Metallicity', 'Temperature'][i], fontsize=14)
+        axs[0].set_title([r'$\log g$', 'C/O', '[M/H]', '$T_{\rm eff}$'][i], fontsize=14)
         sns.histplot(data=residual_train_list, ax=axs[0], label='train', alpha=0.7, bins=19,
                      log_scale=False, stat='percent', legend=True, linewidth=0)
         sns.histplot(data=residual_test_list, label='test', ax=axs[0], alpha=0.3, bins=19,
@@ -176,7 +186,7 @@ def regression_report(trained_model,
 
         f.tight_layout()
         target_name = ['Gravity', 'C_O_ratio', 'Metallicity', 'Temperature'][i]
-        # plt.savefig(f'../outputs/figures/regression_report_{target_name}.pdf', format='pdf')
+        plt.savefig(f'../outputs/figures/regression_report_{target_name}.pdf', format='pdf')
         plt.show()
 
 def filter_dataset_range(dataset, filter_params):
@@ -505,7 +515,8 @@ def filter_dataframe(training_datasets, predicted_targets_dic):
     return nearest_value_list, filtered_df
 
 
-def boxplot_hist(data,
+def boxplot_hist_old(object_name,
+                 data,
                  x_label,
                  xy_loc):
 
@@ -529,15 +540,77 @@ def boxplot_hist(data,
 
     mean = np.round(np.mean(data),2)
     std = np.round(np.std(data),2)
-    plt.annotate(f'{x_label}='+str(np.round(mean,2))+'$\pm$'+str(np.round(std,2)), fontsize=11,
+    plt.annotate(f'{x_label}='+str(np.round(mean,2))+'$\pm$'+str(np.round(std,2)), fontsize=13,
                  xy=(xy_loc[0], xy_loc[1]), xycoords='axes fraction')
 
-    plt.xlabel(x_label, fontsize = 12)
+    plt.xlabel(x_label, fontsize = 14)
+    plt.ylabel(fontsize = 14)
     if x_label == 'C/O':
         x_label = 'c_o_ratio'
     if x_label == '[M/H]':
         x_label = 'metallicity'
-    # plt.savefig(f'../outputs/figures/boxplot_hist_{x_label}.pdf', format='pdf')
+    if x_label == '$T_{\rm eff}$':
+        x_label = 'Teff'
+    if x_label == '$\log g$':
+        x_label = 'logg'
+    plt.tight_layout()
+    plt.savefig(f'../../../outputs/figures/{object_name}__boxplot_hist_{x_label}.pdf', format='pdf')
+
+    plt.show()
+
+
+import seaborn as sns
+import matplotlib.pyplot as plt
+import numpy as np
+
+
+def boxplot_hist(object_name,
+                 data,
+                 x_label,
+                 xy_loc):
[…]
