### 2024-02-27 2de0702
update Predictor module [the recent one] over the writing the ApJ paper on Jan-Feb
