### 2024-02-29 9cd9fc0
add summary to the JOSS draft
