### 2024-02-29 7e56676
update JOSS bib
