+++ publications/joss/paper.bib
@@ -292,6 +292,171 @@ author={
  pages={2825--2830},
  year={2011}
 }
+@ARTICLE{Irwin2008,
+       author = {{Irwin}, P.~G.~J. and {Teanby}, N.~A. and {de Kok}, R. and {Fletcher}, L.~N. and {Howett}, C.~J.~A. and {Tsang}, C.~C.~C. and {Wilson}, C.~F. and {Calcutt}, S.~B. and {Nixon}, C.~A. and {Parrish}, P.~D.},
+        title = "{The NEMESIS planetary atmosphere radiative transfer and retrieval tool}",
+      journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+         year = 2008,
+        month = apr,
+       volume = {109},
+        pages = {1136-1150},
+          doi = {10.1016/j.jqsrt.2007.11.006},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2008JQSRT.109.1136I},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@ARTICLE{Cubillos2021,
+       author = {{Cubillos}, Patricio E. and {Blecic}, Jasmina},
+        title = "{The PYRAT BAY framework for exoplanet atmospheric modelling: a population study of Hubble/WFC3 transmission spectra}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {radiative transfer, methods: statistical, planets and satellites: atmosphere, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2021,
+        month = aug,
+       volume = {505},
+       number = {2},
+        pages = {2675-2702},
+          doi = {10.1093/mnras/stab1405},
+archivePrefix = {arXiv},
+       eprint = {2105.05598},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021MNRAS.505.2675C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Waldmann2015,
+       author = {{Waldmann}, I.~P. and {Tinetti}, G. and {Rocchetto}, M. and {Barton}, E.~J. and {Yurchenko}, S.~N. and {Tennyson}, J.},
+        title = "{Tau-REx I: A Next Generation Retrieval Code for Exoplanetary Atmospheres}",
+      journal = {Astrophysical Journal},
+     keywords = {methods: data analysis, methods: statistical, radiative transfer, techniques: spectroscopic, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2015,
+        month = apr,
+       volume = {802},
+       number = {2},
+          eid = {107},
+        pages = {107},
+          doi = {10.1088/0004-637X/802/2/107},
+archivePrefix = {arXiv},
+       eprint = {1409.2312},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015ApJ...802..107W},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Zhang2019,
+       author = {{Zhang}, Michael and {Chachan}, Yayaati and {Kempton}, Eliza M. -R. and {Knutson}, Heather A.},
+        title = "{Forward Modeling and Retrievals with PLATON, a Fast Open-source Tool}",
+      journal = {Publications of the Astronomical Society of the Pacific},
+     keywords = {Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2019,
+        month = mar,
[…]
