### 2024-02-29 f4c8a52
add these sections: CNN, dataset, compare to similar packages, MODIFY functionality
