+++ publications/joss/paper.md
@@ -61,8 +61,6 @@ Python package developed to perform three main tasks:
 
 
 
-
-
 # Statement of Need
 
 We are on in a new era of space exploration, thanks to advancements in ground- and space-based 
@@ -87,9 +85,6 @@ for interpreting observational data captured by telescopes.
 
 
 
-
-
-
 # Functionality and Key Features
 `TelescopeML` is a Python package comprising a series of modules, each equipped with specialized machine learning and 
 statistical capabilities for conducting Convolutional Neural Networks (CNN) or Machine Learning (ML) training on datasets 
@@ -97,21 +92,23 @@ captured from the atmospheres of extrasolar planets and brown dwarfs. The tasks 
 outlined below and visualized in following Figure:
 
 - **DataMaster module**: Performs various tasks to process the datasets, including:
-    - Preparing inputs and outputs
-    - Splitting the dataset into training, validation, and test sets
-    - Scaling/normalizing the data
-    - Visualizing the data
-    - Conducting feature engineering
+    - Load the training dataset (i.e., atmospheric fluxes) in CSV format
+    - Split the dataset into training, validation, and test sets to pass it to the CNN model
+    - Scale/normalize the dataset column-wise or row-wise
+    - Visualize the training sets in each of the processing step for more insights 
+    - Perform feature engineering by extracting the Min and Max values from each flux to improve the ML training performance
+
   
 - **DeepTrainer module**: Utilizes different methods/packages such as TensorFlow to:
-  - Build Convolutional Neural Networks (CNNs) model using the training examples
-  - Utilize tuned hyperparameters
-  - Fit/train the ML models
+  - Load the processed dataset from the **DataMaster** module
+  - Build Convolutional Neural Networks (CNNs) model using the tuned hyperparameters
+  - Fit/train the CNN models given the epochs, learning rate, and other parameters
   - Visualize the loss and training history, as well as the trained model's performance
   
 - **Predictor module**: Implements the following tasks to predict atmospheric parameters: 
-  - Processes and predicts the observational datasets 
-  - Deploys the trained ML/CNNs model to predict atmospheric parameters
+  - Perform Scale/normalize processes on the observational fluxes
+  - Deploys the trained CNNs model 
+  - Predict atmospheric parameters, i.e., effective temperature, gravity, carbon-to-oxygen ratio, and metallicity 
   - Visualizes the processed observational dataset and the uncertainty in the predicted results
   
 - **StatVisAnalyzer module**: Provides a set of functions to perform the following tasks: 
@@ -120,8 +117,33 @@ outlined below and visualized in following Figure:
   - Calculates confidence intervals and standard errors
   - Functions to visualize the datasets, including scatter plots, histograms, boxplots
 
+
+
+# Details on the synthetic dataset
+
+The training dataset (or synthetic spectra) in this study is computed using the open-source atmospheric radiative 
+transfer Python package, [`PICASO`](https://natashabatalha.github.io/picaso/) [e.g., @batalha2019picaso], based on the 
+`Sonora-Bobcat` model grid generated for cloudless brown dwarf atmospheres by [@marley2021sonora]. This set encompasses 
+30,888 synthetic spectra, each including 104 wavelengths (i.e., 0.897, 0.906, ..., 2.512 μm) and their corresponding flux 
+values. Each of these spectra has four output variables attached to it: effective temperature, gravity, carbon-to-oxygen ratio, 
[…]
