### 2024-03-01 fb87428
update JOSS bib - DOIs
