+++ publications/joss/paper.bib
@@ -7,7 +7,7 @@ volume = {151},
 pages = {107398},
 year = {2021},
 issn = {0888-3270},
-doi = {https://doi.org/10.1016/j.ymssp.2020.107398},
+doi = {10.1016/j.ymssp.2020.107398},
 url = {https://www.sciencedirect.com/science/article/pii/S0888327020307846},
 author = {Serkan Kiranyaz and Onur Avci and Osama Abdeljaber and Turker Ince and Moncef Gabbouj and Daniel J. Inman},
 keywords = {Artificial Neural Networks, Machine learning, Deep learning, Convolutional neural networks, Structural health monitoring, Condition monitoring, Arrhythmia detection and identification, Fault detection, Structural damage detection},
@@ -379,6 +379,8 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+
+
 @ARTICLE{Line2013,
        author = {{Line}, Michael R. and {Wolf}, Aaron S. and {Zhang}, Xi and {Knutson}, Heather and {Kammer}, Joshua A. and {Ellison}, Elias and {Deroo}, Pieter and {Crisp}, Dave and {Yung}, Yuk L.},
         title = "{A Systematic Retrieval Analysis of Secondary Eclipse Spectra. I. A Comparison of Atmospheric Retrieval Techniques}",
@@ -525,6 +527,7 @@ archivePrefix = {arXiv},
   number={1},
   pages={70},
   year={2019},
+  doi = {10.3847/1538-4357/ab1b51},
   publisher={IOP Publishing}
 }
 
