### 2024-03-01 #6346 @EhsanGharibNezhad
Hi Oliver @oparisot:

To address your comments, I added the following sections:
-  _Summary_
-  _Details on the synthetic dataset_
-  _Details on the CNN methodology for Multi-output Regression problem_ to briefly talk about the CNN method
-  _Similar Tools_

I also modified the _Functionality and Key Features_ section to clearly list the tasks. Besides, I updated the DOIs suggested earlier. Please let me know if it still looks vague. Thank you!
