### 2024-03-20 b333d4c
add plot_model_loss_plt function
