+++ TelescopeML/StatVisAnalyzer.py
@@ -11,7 +11,7 @@ from scipy.interpolate import interp1d
 # import pickle as pk
 
 from scipy import stats
-from sklearn.metrics import r2_score, mean_squared_error
+from sklearn.metrics import r2_score, mean_squared_error,  mean_absolute_error
 from scipy.interpolate import RegularGridInterpolator
 from scipy.stats import chi2
 import os
@@ -127,6 +127,15 @@ def regression_report(trained_model,
         r2_score_train = r2_score(y_pred_train, y_act_train)
         r2_score_test = r2_score(y_pred_test, y_act_test)
 
+        # Calculate  MSE scores
+        mse_score_train = mean_squared_error(y_pred_train, y_act_train)
+        mse_score_test = mean_squared_error(y_pred_test, y_act_test)
+        print('MSE ', "{:.2e}".format(mse_score_train) , "{:.2e}".format(mse_score_test) )
+
+        mae_score_train = mean_absolute_error(y_pred_train, y_act_train)
+        mae_score_test = mean_absolute_error(y_pred_test, y_act_test)
+        print('MAE ', "{:.2e}".format(mae_score_train) , "{:.2e}".format(mae_score_test) )
+
         # Calculate  RMSE scores
         rmse_score_train = np.sqrt(mean_squared_error(y_pred_train, y_act_train))
         rmse_score_test = np.sqrt(mean_squared_error(y_pred_test, y_act_test))
@@ -139,14 +148,14 @@ def regression_report(trained_model,
         if __print_results__:
             print('\n\n----------------------- Test ------------------------')
             print('R2: {:2.2f} \t  RMSE: {:2.2f} \t Mean+/-STD: {:2.2f}+/-{:2.2f}'.format(
-                r2_score_test, rmse_score_train, mean_test, std_test))
+                r2_score_test, rmse_score_train, mean_test, std_test, mse_score_test))
 
             print('\n----------------------- Train ------------------------')
             print('R2: {:2.2f} \t  RMSE: {:2.2f} \t Mean+/-STD: {:2.2f}+/-{:2.2f}'.format(
                 r2_score_train, rmse_score_test, mean_train, std_train))
 
         # Plot histograms of residuals
-        axs[0].set_title(['Gravity', 'C_O_ratio', 'Metallicity', 'Temperature'][i], fontsize=14)
+        axs[0].set_title(['$\log g$', 'C/O', '[M/H]', r'$T_{\rm eff}$'][i], fontsize=14)
         sns.histplot(data=residual_train_list, ax=axs[0], label='train', alpha=0.7, bins=19,
                      log_scale=False, stat='percent', legend=True, linewidth=0)
         sns.histplot(data=residual_test_list, label='test', ax=axs[0], alpha=0.3, bins=19,
@@ -155,6 +164,7 @@ def regression_report(trained_model,
         axs[0].set_ylim((1e-1, 100))
         axs[0].set_yscale('log')
         axs[0].set_ylabel('Probability %', fontsize=12)
+        axs[0].set_xlabel('Residual', fontsize=12)
 
         # Plot scatter figures of predicted vs actual values
         sns.scatterplot(y=y_pred_train, x=y_act_train, label='train', ax=axs[1], alpha=0.7, legend=False)
@@ -163,20 +173,27 @@ def regression_report(trained_model,
         axs[1].set_xlabel('Actual value', fontsize=12)
 
         # Add annotations for skewness and R-squared scores
[…]
