### 2024-03-24 #6346 @plaplant
@EhsanGharibNezhad for comments on the submission, we leave that up to the reviewers' preference. Some will make comments in this review thread, and others will make issues on the software repository and reference them as part of this thread, especially if the comments are more involved or there are several issues that need to be addressed. Either way, at least a reference should appear in this thread.

@oparisot thanks very much for your initial review! When you get a chance, please provide any additional feedback you have, either as a comment on this thread or by making an issue on [the software repository](https://github.com/EhsanGharibNezhad/TelescopeML). Thanks!

@mwalmsley when you get a chance, please make the following comment in this thread:
Then please start going through the review elements contained in the checklist. Let me know if you have any questions!
