### 2024-04-05 #6346 @oparisot
The software is easy to use by following the instructions, and the use-cases are well explained. I'm in favour of accepting the paper, if guidelines (even slight ones) for contribution / support are added in the git repository.

Olivier
