### 2024-04-05 #6346 @EhsanGharibNezhad
> Hi @EhsanGharibNezhad,
>
> Thank you for the update of the paper. Now, I'm testing the software -- I will give you my feedback.
>
> Ideas to improve the repository: can you add a short 'contribution guide' (ex: https://contributing.md/example/)?
>
> Olivier

Hi Oliver @oparisot, This is a great comment. Thanks for sharing this link. I added two docs to the main branch to address that:

1. **CODE_OF_CONDUCT.md**: https://github.com/EhsanGharibNezhad/TelescopeML/blob/main/CODE_OF_CONDUCT.md
2. **CONTRIBUTING.md**: https://github.com/EhsanGharibNezhad/TelescopeML/blob/main/CONTRIBUTING.md
