### 2024-04-08 22b6265
update wavelength list str
