+++ TelescopeML/DataMaster.py
@@ -38,7 +38,7 @@ tf.get_logger().setLevel('ERROR')
 
 # ===============================================================================
 # ==================                                           ==================
-# ==================            Train CNN Regression           ==================
+# ==================            Data Processor                 ==================
 # ==================                                           ==================
 # ===============================================================================
 
@@ -58,23 +58,17 @@ class DataProcessor:
         Flux arrays (input data).
     wavelength_names : List[str]
         Name of wavelength in micron.
+    wavelength_values : np.ndarray
+        wavelength array in micron.
     output_values : np.ndarray
-        Target variable array (e.g., Temperature, Gravity, Carbon_to_Oxygen, Metallicity).
-    output_names : str
-        Name of the target variable.
-    is_hyperparam_tuned : str
-        Indicates whether hyperparameters are tuned or not ('yes' or 'no').
-    param_grid : dict, optional
-        ML hyperparameters to be tuned (used if is_hyperparam_tuned = 'yes').
+        output variable array (e.g., Temperature, Gravity, Carbon_to_Oxygen, Metallicity).
+    output_names : List[str]
+        Name of the output variable.
     spectral_resolution : int, optional
         Resolution of the synthetic spectra used to generate the dataset.
-    feature_improvement_method : str
-        Method used for feature improvement ('no', 'pca', 'RFE').
-    augmentation_method : str
-        Indicates if an augmented dataset is used ('no' or method name).
-    ml_model : BaseEstimator, optional
+    trained_ML_model : BaseEstimator, optional
         ML model object from sklearn package.
-    ml_model_str : str, optional
+    trained_ML_model_name : str, optional
         Name of the ML model.
     ml_method : str, optional
         Machine learning method ('regression' or 'classification').
@@ -84,44 +78,36 @@ class DataProcessor:
             self,
             flux_values: Union[np.ndarray] = None,
             wavelength_names: Union[List[str]] = None,
+            wavelength_values: Union[np.ndarray] = None,
             output_values: Union[np.ndarray] = None,
             output_names: Union[str] = None,
-            is_tuned: str = 'no',
-            param_grid: Union[None, Dict] = None,
             spectral_resolution: Union[None, int] = None,
-            is_feature_improved: str = 'no',
-            is_augmented: str = 'no',
-            ml_model: Union[None, BaseEstimator] = None,
-            ml_model_str: Union[None, str] = None,
+            trained_ML_model: Union[None, BaseEstimator] = None,
+            trained_ML_model_name: Union[None, str] = None,
             ml_method: str = 'regression',
     ):
 
         self.flux_values = flux_values
         self.wavelength_names = wavelength_names
+        self.wavelength_values = wavelength_values
         self.output_values = output_values
         self.output_names = output_names
-        self.is_tuned = is_tuned
-        self.param_grid = param_grid
         self.spectral_resolution = spectral_resolution
-        self.is_feature_improved = is_feature_improved
-        self.is_augmented = is_augmented
-        self.ml_model = ml_model
-        self.ml_model_str = ml_model_str
+        self.trained_ML_model = trained_ML_model
+        self.trained_ML_model_name = trained_ML_model_name
         self.ml_method = ml_method
-        self.LoadSave = LoadSave(ml_model_str,
+        self.LoadSave = LoadSave(trained_ML_model_name,
                                  ml_method,
-                                 is_feature_improved,
-                                 is_augmented,
-                                 is_tuned)
+                                 )
 
     def split_train_test(self, test_size=0.1):
         """
-        Split the loaded set into train and test sets
+        Split the loaded dataset into train and test sets
 
         Parameters
         ----------
         test_size : float
-            The proportion of the dataset to include in the test split.
+            The proportion of the dataset to include in the test split (default = 0.1).
 
         Returns
         -------
[…]
