### 2024-04-09 853004d
Update StatVisAnalyzer.py
