+++ TelescopeML/StatVisAnalyzer.py
@@ -28,6 +28,9 @@ from bokeh.plotting import output_notebook
 from bokeh.plotting import figure, show
 from bokeh.models import ColumnDataSource
 
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.ticker import MaxNLocator
 
 def print_results_fun(targets, print_title=None):
     """
@@ -259,7 +262,7 @@ def interpolate_df(dataset,
         print(my_list_g, my_list_c_o, my_list_T, my_list_met)
 
     df_interpolated_ = pd.DataFrame(columns=df_to_interpolate.drop(
-        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity', 'is_augmented']).columns)
+        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity']).columns)
 
     df_interpolated_all = pd.DataFrame(columns=df_to_interpolate.columns)
 
@@ -278,13 +281,13 @@ def interpolate_df(dataset,
                         drop=True)  # .drop_duplicates(subset=['gravity', 'temperature', 'c_o_ratio', 'metallicity'])
 
                     data = df_to_interpolate_.drop(
-                        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity', 'is_augmented'])
+                        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity'])
 
                     y = df_to_interpolate_['c_o_ratio'].to_numpy()
 
                     column_grid = data.columns.to_numpy().astype(float)
                     values = df_to_interpolate_.drop(
-                        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity', 'is_augmented']).to_numpy()
+                        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity']).to_numpy()
 
                     interp_func = RegularGridInterpolator((y, column_grid), values)
 
@@ -302,7 +305,7 @@ def interpolate_df(dataset,
                     df_interpolated_['temperature'] = temp
                     df_interpolated_['metallicity'] = met
                     df_interpolated_['gravity'] = grav
-                    df_interpolated_['is_augmented'] = 'no'
+                    # df_interpolated_['is_augmented'] = 'no'
 
                     df_interpolated_all = pd.concat([df_interpolated_, df_interpolated_all], ignore_index=True)
 
@@ -317,7 +320,7 @@ def interpolate_df(dataset,
     df_interpolated_all2.drop_duplicates(inplace=True)
 
     df_interpolated_ = pd.DataFrame(columns=df_interpolated_all2.drop(
-        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity', 'is_augmented']).columns)
+        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity',]).columns)
 
     for c_o in my_list_c_o:
         for temp in my_list_T:
@@ -333,13 +336,13 @@ def interpolate_df(dataset,
                         drop=True)  # .drop_duplicates(subset=['gravity', 'temperature', 'c_o_ratio', 'metallicity'])
 
                     data = df_to_interpolate_.drop(
-                        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity', 'is_augmented'])
+                        columns=['gravity', 'temperature', 'c_o_ratio', 'metallicity', ])
 
                     y = df_to_interpolate_['metallicity'].to_numpy()
 
                     column_grid = data.columns.to_numpy().astype(float)
                     values = df_to_interpolate_.drop(
[…]
