+++ TelescopeML/Predictor.py
@@ -6,7 +6,7 @@ from .StatVisAnalyzer import boxplot_hist, plot_spectra_errorbar, \
     plot_pred_vs_obs_errorbar
 from .StatVisAnalyzer import interpolate_df, print_results_fun
 from .StatVisAnalyzer import replace_zeros_with_mean, calculate_confidence_intervals_std_df, \
-    plot_pred_vs_obs_errorbar_stat
+    plot_pred_vs_obs_errorbar_stat_bokeh, plot_pred_vs_obs_errorbar_stat_matplotlib
 
 # ======= Import Python libraries ========================================
 
@@ -457,9 +457,9 @@ class ObserveParameterPredictor:
             __plot_randomly_generated_spectra__=False,
             __plot_histogram__=False,
             __plot_boxplot_hist__=False,
-            # __plot_predicted_vs_observed__=False,
             __plot_pred_vs_obs_errorbar__=False,
-            __plot_pred_vs_obs_errorbar_stat__=False,
+            __plot_pred_vs_obs_errorbar_stat_bokeh__=False,
+            __plot_pred_vs_obs_errorbar_stat_matplotlib__=False,
             __calculate_confidence_intervals_std_df__=False,
             ):
         """
@@ -480,7 +480,7 @@ class ObserveParameterPredictor:
             True or False.
         __plot_pred_vs_obs_errorbar__ : bool
             True or False.
-        __plot_pred_vs_obs_errorbar_stat__ : bool
+        __plot_pred_vs_obs_errorbar_stat_bokeh__ : bool
             True or False.
         __calculate_confidence_intervals_std_df__ : bool
             True or False.
@@ -662,8 +662,8 @@ class ObserveParameterPredictor:
                 __plot_calculate_confidence_intervals_std_df__=False,
             )
 
-        if __plot_pred_vs_obs_errorbar_stat__:
-            plot_pred_vs_obs_errorbar_stat(
+        if __plot_pred_vs_obs_errorbar_stat_bokeh__:
+            plot_pred_vs_obs_errorbar_stat_bokeh(
                 stat_df=self.confidence_intervals_std_df,
                 confidence_level=0.95,
                 object_name=self.object_name,
@@ -676,6 +676,19 @@ class ObserveParameterPredictor:
                 radius = self.bd_literature_dic['bd_radius_Rjup'],
                 __print_results__=False,
             )
+
+        if __plot_pred_vs_obs_errorbar_stat_matplotlib__:
+            plot_pred_vs_obs_errorbar_stat_matplotlib(stat_df=self.confidence_intervals_std_df,
+                                                      confidence_level=0.95,
+                                                      object_name=self.object_name,
+                                                      x_obs=self.obs_data_df['wl'],
+                                                      y_obs=self.obs_data_df['Fnu_obs_absolute'],
+                                                      y_obs_err=self.obs_data_df['Fnu_obs_absolute_err'],
+                                                      training_datasets=self.training_dataset_df,
+                                                      x_pred=self.wl_synthetic,
+                                                      predicted_targets_dic=self.dic_random_pred_mean,
+                                                      radius=self.bd_literature_dic['bd_radius_Rjup'],
+                                                      __print_results__=False)
 
 
 
