### 2024-04-09 2793de3
update StatVisAnalyzer.py
