+++ TelescopeML/StatVisAnalyzer.py
@@ -15,6 +15,7 @@ from sklearn.metrics import r2_score, mean_squared_error
 from scipy.interpolate import RegularGridInterpolator
 from scipy.stats import chi2
 import os
+from matplotlib.ticker import AutoMinorLocator
 
 import pprint
 
@@ -495,6 +496,7 @@ def interpolate_df(dataset,
     return df_interpolated_final
 
 
+# def filter_dataframe(training_datasets, predicted_targets_dic):
 def filter_dataframe(training_datasets, predicted_targets_dic):
     nearest_value_list = []
     filtered_df = training_datasets.copy()
@@ -712,7 +714,7 @@ def plot_pred_vs_obs_errorbar(object_name,
 
     show(p)
 
-def plot_pred_vs_obs_errorbar_stat(  stat_df,
+def plot_pred_vs_obs_errorbar_stat_bokeh(  stat_df,
                                 confidence_level,
                                 object_name,
                                 x_obs,
@@ -772,7 +774,7 @@ def plot_pred_vs_obs_errorbar_stat(  stat_df,
     # Create ML figure
     p = figure(
         title=object_name+': Observational vs. ML Predicted Spectra'+' [𝛘2='+str(chi2_stat)+', p-value='+ str(p_value)+']',
-        x_axis_label='Features (Wavelength [μm])',
+        x_axis_label='Wavelength [μm]',
         y_axis_label='Absolute Flux (F𝜈) [erg/s/cm2/Hz]',
         y_axis_type="log",
         width=1000,
@@ -915,7 +917,7 @@ def calculate_confidence_intervals_std_df(dataset_df,
         x = np.round(df3.columns, 2)
         p = figure(
             title='Mean with Confidence Intervals',
-            x_axis_label='Features (Wavelength [μm])',
+            x_axis_label='Wavelength [μm]',
             y_axis_label='Flux (F𝜈) [erg/s/cm2/Hz]',
             y_axis_type="log",
             width=1000,
@@ -1376,7 +1378,7 @@ def plot_filtered_dataframe_notUsed(dataset, filter_bounds, feature_to_plot, tit
                         color=colors[i], alpha=0.7)
 
     # print(filtered_data.T[col][:4].values[0])
-    ax.set_xlabel('Features (Wavelength [$\mu$m])')
+    ax.set_xlabel('Wavelength [$\mu$m]')
     ax.set_ylabel(r'F$_{\nu}$  [erg/cm$^2$/s/Hz]')
     dict_features = {'temperature': 'Effective Temperature', 'gravity': 'Gravity', 'metallicity': 'Metallicity',
                      'c_o_ratio': 'Carbon-to-oxygen ratio'}
@@ -1579,4 +1581,266 @@ def plot_tricontour_chi2_radius(tuned_ML_R_param_df,
                         tuned_ML_R_param_df['bd_name'] + '_TunedRadius_' + target
                         + '_counterplot.pdf',
                         format='pdf', bbox_inches='tight')
+        plt.show()
+
+
+def plot_pred_vs_obs_errorbar_stat_matplotlib(  stat_df,
+                                confidence_level,
+                                object_name,
+                                x_obs,
+                                y_obs,
+                                y_obs_err,
+                                training_datasets,
+                                x_pred,
+                                predicted_targets_dic,
+                                radius,
+                                __print_results__ = False):
+    """
+    Plot observed spectra with error bars and predicted spectra with confidence intervals.
+
+    Parameters
+    ----------
+    stat_df : DataFrame
[…]
