### 2024-04-09 0cd334c
update some names/funcs inside predictor
