+++ TelescopeML/StatVisAnalyzer.py
@@ -57,131 +57,6 @@ def print_results_fun(targets, print_title=None):
 
 
 
-def regression_report(trained_model,
-                      trained_data_processor,
-                      Xtrain, Xtest, ytrain, ytest,
-                      target_i,
-                      xy_top=None, xy_bottom=None, __print_results__= False):
-    """
-    Generate a regression report for the trained ML/CNN model.
-
-    Parameters
-    -----------
-    trained_model : object
-        Trained regression model.
-    Xtrain : array
-        Training set.
-    Xtest : array
-        Test set.
-    ytrain : array
-        Training target set.
-    ytest : array
-        Test target set.
-    target_i : int
-        Index of the target variable to analyze.
-    xy_top : list, optional
-        Coordinates for annotations in the top plot. Defaults to [0.55, 0.85].
-    xy_bottom : list, optional
-        Coordinates for annotations in the bottom plot. Defaults to [0.05, 0.8].
-    __print_results__ : bool, optional
-        True or False.
-    """
-
-    # Apply the trained ML model on the train set to predict the targets
-    if xy_bottom is None:
-        xy_bottom = [0.05, 0.8]
-    if xy_top is None:
-        xy_top = [0.55, 0.85]
-    y_pred_train = np.array(trained_model.predict(Xtrain))[:, :, 0].T
-    y_pred_train_list = trained_data_processor.standardize_y_ColumnWise.inverse_transform(y_pred_train)
-    y_pred_train_list[:, 3] = 10 ** y_pred_train_list[:, 3]
-
-    y_act_train_list = trained_data_processor.standardize_y_ColumnWise.inverse_transform(ytrain)
-    y_act_train_list[:, 3] = 10 ** y_act_train_list[:, 3]
-
-    # Apply the trained ML model on the test set to predict the targets
-    y_pred_test = np.array(trained_model.predict(Xtest))[:, :, 0].T
-    y_pred_test_list = trained_data_processor.standardize_y_ColumnWise.inverse_transform(y_pred_test)
-    y_pred_test_list[:, 3] = 10 ** y_pred_test_list[:, 3]
-
-    y_act_test_list = trained_data_processor.standardize_y_ColumnWise.inverse_transform(ytest)
-    y_act_test_list[:, 3] = 10 ** y_act_test_list[:, 3]
-
-    for i in range(0, target_i):
-        y_pred_train = y_pred_train_list[:, i]
-        y_act_train = y_act_train_list[:, i]
-        y_pred_test = y_pred_test_list[:, i]
-        y_act_test = y_act_test_list[:, i]
-
-        # Calculate the residual (Predicted - Actual)
-        residual_train_list = y_pred_train - y_act_train
-        residual_test_list = y_pred_test - y_act_test
-
-        # Calculate mean and standard deviation for residuals
-        mean_test = np.round(np.mean(residual_test_list), 2)
-        std_test = np.round(np.std(residual_test_list), 2)
-        mean_train = np.round(np.mean(residual_train_list), 2)
-        std_train = np.round(np.std(residual_train_list), 2)
-
-        # Calculate skewness for residuals
-        skew_test = stats.skew(residual_test_list)
-        skew_train = stats.skew(residual_train_list)
-
-        # Calculate R-squared scores
-        r2_score_train = r2_score(y_pred_train, y_act_train)
-        r2_score_test = r2_score(y_pred_test, y_act_test)
-
-        # Calculate  RMSE scores
[…]
