+++ TelescopeML/DeepTuner.py
@@ -1,17 +1,17 @@
-import scipy.signal.signaltools
-
-
-def _centered(arr, newsize):
-    # Return the center newsize portion of the array.
-    newsize = np.asarray(newsize)
-    currsize = np.array(arr.shape)
-    startind = (currsize - newsize) // 2
-    endind = startind + newsize
-    myslice = [slice(startind[k], endind[k]) for k in range(len(endind))]
-    return arr[tuple(myslice)]
-
-
-scipy.signal.signaltools._centered = _centered
+# import scipy.signal.signaltools
+#
+#
+# def _centered(arr, newsize):
+#     # Return the center newsize portion of the array.
+#     newsize = np.asarray(newsize)
+#     currsize = np.array(arr.shape)
+#     startind = (currsize - newsize) // 2
+#     endind = startind + newsize
+#     myslice = [slice(startind[k], endind[k]) for k in range(len(endind))]
+#     return arr[tuple(myslice)]
+#
+#
+# scipy.signal.signaltools._centered = _centered
 
 import os
 
@@ -36,7 +36,7 @@ import argparse
 import hpbandster.core.nameserver as hpns
 import hpbandster.core.result as hpres
 
-from hpbandster.optimizers import BOHB as BOHB
+# from hpbandster.optimizers import BOHB as BOHB
 # from hpbandster.examples.commons import MyWorker
 
 # import logging
@@ -371,71 +371,74 @@ class KerasWorker(Worker):
 # Read the `TelescopeML_reference_data` path
 
 import os  # to check the path
-from ConfigSpace.hyperparameters import UniformIntegerHyperparameter, UniformFloatHyperparameter, \
-    CategoricalHyperparameter
 
 __reference_data_path__ = os.getenv("TelescopeML_reference_data")
 __reference_data_path__
 
-# Step 1: Load the dataset
 train_BD = pd.read_csv(os.path.join(__reference_data_path__,
                                     'training_datasets',
-                                    'browndwarf_R100_v4_newWL_v2.csv.bz2'), compression='bz2')
+                                    'browndwarf_R100_v4_newWL_v3.csv.bz2'), compression='bz2')
+train_BD.head(5)
+
+
+output_names = ['gravity', 'temperature', 'c_o_ratio', 'metallicity']
+train_BD[output_names].head()
+
+
+
+wavelength_names = [item for item in train_BD.columns.to_list() if item not in output_names]
+wavelength_names[:5]
+
+
+wavelength_values = [float(item) for item in wavelength_names]
+wavelength_values[:10]
+
 
 wl_synthetic = pd.read_csv(os.path.join(__reference_data_path__,
                                         'training_datasets',
                                         'wl.csv'))
+wl_synthetic.head(3)
 
-## Prepare feature variables (X) and targets (y)
 # to assure we are only training the module with the native non-augmented BD training dataset
-train_BD = train_BD[train_BD['is_augmented'].isin(['no'])]
+train_BD = train_BD
 
-target_features = ['gravity', 'temperature', 'c_o_ratio', 'metallicity']
-training_features_labels = [item for item in train_BD.columns.to_list() if
-                            item not in target_features + ['is_augmented']]
 
 # Training feature variables
 X = train_BD.drop(
     columns=['gravity',
              'temperature',
              'c_o_ratio',
-             'metallicity',
-             'is_augmented'])  # .astype(np.float32)
+             'metallicity'])#.astype(np.float32)
+
 
 # Target/Output feature variables
-y = train_BD[['gravity', 'c_o_ratio', 'metallicity', 'temperature', ]]  # .astype(np.float32)
+y = train_BD[['gravity', 'c_o_ratio', 'metallicity', 'temperature', ]]#.astype(np.float32)
 
[…]
