+++ TelescopeML/Predictor.py
@@ -189,7 +189,7 @@ class ObserveParameterPredictor:
 
         Return
         -------
-        Fnu_obs , Fnu_obs_err, Fnu_obs_absolute, Fnu_obs_absolute_err
+        Fnu_obs , Fnu_obs_err, Fnu_obs_TOA, Fnu_obs_TOA_err
 
         Example
         -------
@@ -215,13 +215,13 @@ class ObserveParameterPredictor:
                  Flam_errors = self.F_lambda_obs_err,
                  wavelengths = wl_obs,)  # Convert F_lambda_obs to Fnu_obs
 
-        Fnu_obs_absolute, Fnu_obs_absolute_err = self.Fnu_to_Fnu_abs(
+        Fnu_obs_TOA, Fnu_obs_TOA_err = self.Fnu_to_Fnu_abs(
                  Fnu_values = Fnu_obs,
                  Fnu_errors = Fnu_obs_err,
                  bd_literature_dic = bd_literature_dic)
 
 
-        return Fnu_obs , Fnu_obs_err, Fnu_obs_absolute, Fnu_obs_absolute_err
+        return Fnu_obs , Fnu_obs_err, Fnu_obs_TOA, Fnu_obs_TOA_err
 
 
     def Fnu_to_Fnu_abs(self,
@@ -231,7 +231,7 @@ class ObserveParameterPredictor:
                        __plot__ = False
                        ):
 
-        # Calculate the absolute Fnu_obs values
+        # Calculate the TOA Fnu_obs values
         bd_literature_dic = self.bd_literature_dic if bd_literature_dic is None else bd_literature_dic
 
         Fnu_values_abs = Fnu_values * (bd_literature_dic['bd_distance_pc'] * ((u.pc).to(u.jupiterRad)) / (
@@ -289,7 +289,7 @@ class ObserveParameterPredictor:
         return Fnu_values, Fnu_errors
 
     def flux_interpolated(self,
-                          Fnu_obs_absolute,
+                          Fnu_obs_TOA,
                           interpolated_wl = None,
                           __print_results__=False,
                           __plot_spectra_errorbar__=False,
@@ -308,28 +308,28 @@ class ObserveParameterPredictor:
 
         Return
         -------
-        Fnu_obs_absolute_intd, Fnu_obs_absolute_intd_df
+        Fnu_obs_TOA_intd, Fnu_obs_TOA_intd_df
         """
-        Fnu_obs_absolute = self.Fnu_obs_absolute if Fnu_obs_absolute is None else Fnu_obs_absolute
+        Fnu_obs_TOA = self.Fnu_obs_TOA if Fnu_obs_TOA is None else Fnu_obs_TOA
         interpolated_wl = np.sort(self.wl_synthetic.wl.values) if interpolated_wl is None else interpolated_wl
 
         if __use_spectres__:
-            Fnu_obs_absolute_intd = spectres.spectres(interpolated_wl,
+            Fnu_obs_TOA_intd = spectres.spectres(interpolated_wl,
                                                            np.float64(self.wl_obs),
-                                                           np.float64(Fnu_obs_absolute))
+                                                           np.float64(Fnu_obs_TOA))
         else:
-            flux_intd = pchip(self.wl_obs, self.Fnu_obs_absolute)
-            Fnu_obs_absolute_intd = flux_intd(interpolated_wl)
+            flux_intd = pchip(self.wl_obs, self.Fnu_obs_TOA)
+            Fnu_obs_TOA_intd = flux_intd(interpolated_wl)
 
-        Fnu_obs_absolute_intd_df = pd.DataFrame(list(Fnu_obs_absolute_intd),
+        Fnu_obs_TOA_intd_df = pd.DataFrame(list(Fnu_obs_TOA_intd),
[…]
