### 2024-04-11 4f3a26a
update func in StatVisAnalyzer.py
