+++ TelescopeML/StatVisAnalyzer.py
@@ -544,9 +544,9 @@ def plot_pred_vs_obs_errorbar(object_name,
     source = ColumnDataSource(data=dict(x=x_obs, y=y_obs, upper=upper, lower=lower))
 
     # Create the Observational figure
-    p = figure(title=f"{object_name}: Calibrated Observational VS. Predicted Spectra",
+    p = figure(title=f"{object_name}: Observational vs. ML Predicted Spectra",
                x_axis_label="Wavelength [𝜇m]",
-               y_axis_label="Flux (F𝜈) [erg/s/cm2/Hz]",
+               y_axis_label="TOA Flux (F𝜈) [erg/s/cm2/Hz]",
                width=800, height=300,
                y_axis_type="log",
                tools="pan,wheel_zoom,box_zoom,reset")
@@ -1109,14 +1109,6 @@ def find_closest_chi_square(df, chi_square_statistic_list):
 
     return closest_chi_square, closest_p_value
 
-# Example usage with df = 103 and chi_square_list containing chi-square statistics
-# df_value = 103
-# chi_square_list = [93, 32,  150.456789123, 120.789123456]  # Replace with actual chi-square statistics
-
-
-
-# check_chi_square(103, chi_square_list)
-
 
 def plot_scatter_x_y (x, y,
                         plot_title="Scatter Plot",
@@ -1279,16 +1271,14 @@ def plot_filtered_dataframe_notUsed(dataset, filter_bounds, feature_to_plot, tit
     plt.show()
 
 
-def plot_model_loss(history=None, title=None):
+def plot_ML_model_loss(trained_ML_model_history=None, title=None):
     """
     Plot the trained model history for all individual target features
     """
-    # from bokeh.plotting import figure, show
-    # from bokeh.models import Legend
 
     # history = self.trained_model_history if history is None else history
     # Define the epochs as a list
-    epochs = list(range(len(history['loss'])))
+    epochs = list(range(len(trained_ML_model_history['loss'])))
 
     # Define colorblind-friendly colors
     colors = ['#d62728', '#ff7f0e', '#2ca02c', '#9467bd', '#8c564b']
@@ -1297,25 +1287,25 @@ def plot_model_loss(history=None, title=None):
     p = figure(title=title, width=1000, height=300, y_axis_type='log', x_axis_label='Epochs', y_axis_label='Loss')
 
     # Add the data lines to the figure with colorblind-friendly colors and increased line width
-    p.line(epochs, history['loss'], line_color=colors[0], line_dash='solid', line_width=2,
+    p.line(epochs, trained_ML_model_history['loss'], line_color=colors[0], line_dash='solid', line_width=2,
            legend_label='Total loss')
-    p.line(epochs, history['val_loss'], line_color=colors[0], line_dash='dotted', line_width=2)
+    p.line(epochs, trained_ML_model_history['val_loss'], line_color=colors[0], line_dash='dotted', line_width=2)
 
-    p.line(epochs, history['output__gravity_loss'], line_color=colors[1], line_dash='solid', line_width=2,
+    p.line(epochs, trained_ML_model_history['output__gravity_loss'], line_color=colors[1], line_dash='solid', line_width=2,
            legend_label='gravity')
-    p.line(epochs, history['val_output__gravity_loss'], line_color=colors[1], line_dash='dotted', line_width=2)
+    p.line(epochs, trained_ML_model_history['val_output__gravity_loss'], line_color=colors[1], line_dash='dotted', line_width=2)
 
[…]
