### 2024-04-19 #6346 @mwalmsley
Hello. I am sorry it has taken me some time to get to this. From a look at the documentation and the paper, this is clearly a substantial piece of scholarly work and will benefit the field.

It's not easy for me to try the software right now because I'm traveling and won't be back for another two weeks. Ping me if you need a faster review than that for any reason and I'll give it a try earlier.

(my one very unhelpful comment is that TelescopeML is a rather ambiguous name. Many astroML projects also use telescopes ;) )
