### 2024-04-21 6156e14
update the save synthetic fig: plot_filtere_data
