### 2024-04-22 e11139e
update StatVisAnalyzer.py
