### 2024-04-22 #6346 @EhsanGharibNezhad
Hi Mike, Thanks for the heads-up. Two weeks should be okay as we are currently pushing a new version of the package and updating the Sphinx tutorials. Regarding the name, I agree that it is pretty general and it can be ambiguous, but this is a long-term program which includes different data analysis problems related to spectral analysis and interpretation of observational data.
