+++ _sources/tutorials.rst.txt
@@ -12,7 +12,6 @@ Exploring the Datasets
 
 
 
-
 Predict Atmospheric Parameters
 -------------------------------
 .. toctree::
@@ -22,14 +21,12 @@ Predict Atmospheric Parameters
 
 
 
-
 Data Processing
 ----------------
 .. toctree::
    :maxdepth: 2
 
    Data Processing  <tutorials/3__Data_Processing.ipynb>
-
 
 
 Train a Regression CNNs Model
