+++ TelescopeML/LIMEXplainer.py
@@ -1,12 +1,15 @@
 import numpy as np
-import copy
 from sklearn.metrics import pairwise_distances
 from sklearn.linear_model import LinearRegression
 
+
 class LIMEXplainer:
-    def __init__(self, model, X_test, instance_id, num_segments=50, num_perturbations=200, kernel_width=0.25, perturb_function=None, wavelengths=None):
+    def __init__(self, model, X_test, instance_id, num_segments=50,
+                 num_perturbations=200, kernel_width=0.25,
+                 perturb_function=None, wavelengths=None):
         """
-        Initializes the LIMEXplainer with the model, test data, instance ID, and configuration for LIME analysis.
+        Initializes the LIMEXplainer with the model, test data, instance ID,
+        and configuration for LIME analysis.
         """
         self.model = model
         self.X_test = X_test
@@ -14,19 +17,24 @@ class LIMEXplainer:
         self.num_segments = num_segments
         self.num_perturbations = num_perturbations
         self.kernel_width = kernel_width
-        self.perturb_function = perturb_function if perturb_function else self.perturb_segment_mean
-        self.wavelengths = np.array(wavelengths) if wavelengths is not None else None
+        self.perturb_function = (perturb_function if perturb_function
+                                 else self.perturb_segment_mean)
+        self.wavelengths = (np.array(wavelengths)
+                            if wavelengths is not None else None)
         self.random_perturbations = self.generate_random_perturbations()
         self.perturb_and_store_spectra()
         self.cosine_distances = self.calc_cosine_similarity()
         self.weights = self.compute_weights_from_similarity()
-        self.surrogate_models, self.importance_coefficients = self.fit_surrogate_models()
+        self.surrogate_models, self.importance_coefficients = (
+            self.fit_surrogate_models()
+        )
         self.top_segments_list = self.analyze()
 
-    # Initialize Instance 
+    # Initialize Instance
     def set_instance_by_id(self, instance_id):
         """
-        Sets the current instance of the spectrum and related data based on the given instance ID.
+        Sets the current instance of the spectrum and related data
+        based on the given instance ID.
 
         Parameters:
         - instance_id (int): The ID of the spectrum instance to set.
@@ -37,11 +45,12 @@ class LIMEXplainer:
 
     def get_spectrum_instance_by_id(self, instance_id):
         """
-        Retrieves and sets a spectrum instance by its ID from the stored test data.
-        
+        Retrieves and sets a spectrum instance by its ID
+        from the stored test data.
+
         Parameters:
         - instance_id (int): The ID of the spectrum instance to retrieve.
-        
+
         Returns:
         - np.ndarray: The spectrum instance.
         """
@@ -49,21 +58,22 @@ class LIMEXplainer:
 
         return self.spectrum_instance
 
-    # Perturbation Functions 
+    # Perturbation Functions
     @staticmethod
     def perturb_segment_mean(spectrum, start_idx, end_idx):
         """
-        Returns a new spectrum array where a segment of the original spectrum is replaced with its own mean value.
+        Returns a new spectrum array where a segment of the original spectrum
+        is replaced with its own mean value.
 
         Parameters:
         - spectrum (np.ndarray): The original spectrum.
         - start_idx (int): The starting index of the segment to be perturbed.
-        - end_idx (int): The ending index (exclusive) of the segment to be perturbed.
+        - end_idx (int): The ending index of the segment to be perturbed.
 
         Returns:
-        - np.ndarray: A new spectrum array with the specified segment perturbed.
+        - np.ndarray: New spectrum array with the specified segment perturbed.
         """
-        if end_idx > start_idx:  
[…]
