### 2024-05-06 c4f7307
edit plot_filtered_spectra
