### 2024-05-06 fe87764
update plot_regression_report
