### 2024-05-09 #6346 @mwalmsley
Hi @EhsanGharibNezhad . I'm working through the installation and found the following error with conda:

```
ResolvePackageNotFound:
  - pyobjc-core=9.2
  - pyobjc-framework-cocoa=9.2
  - libgfortran=5.0.0
```

From a quick look, I think the conda forge version of pyobjc-core is only available on osx (see the installers tags [here](https://anaconda.org/conda-forge/pyobjc-core). Similarly libgfortran is available on all platforms but 5.0.0 only seems to have osx installers [here](https://anaconda.org/conda-forge/libgfortran/files?page=1&version=5.0.0).

Could you please update the .yml? Then I will reinstall and run the notebooks, and then approve :)

Cheers,
Mike
