+++ TelescopeML/StatVisAnalyzer.py
@@ -22,6 +22,8 @@ import pprint
 # ******* Data Visulaization Libraries ****************************
 import seaborn as sns
 import matplotlib.pyplot as plt
+from matplotlib.colors import Normalize
+from matplotlib.cm import ScalarMappable
 
 
 from bokeh.plotting import output_notebook
@@ -1699,325 +1701,328 @@ def plot_regression_report(trained_ML_model,
 
 
 # SpectrumVisualizer for LIMEXplainer Module
-class SpectrumVisualizer:
-    def __init__(self, explainer=None, wavelengths=None):
-        """
-        Initializes the SpectrumVisualizer with
-        an optional LIMEXplainer instance.
+def plot_spectrum_instance(explainer):
+    """
+    Plots the selected spectrum.
 
-        Parameters:
-        - explainer (LIMEXplainer, optional): An instance of LIMEXplainer class
-          for accessing spectrum, segments, and perturbations.
-        """
-        self.explainer = explainer
-        if wavelengths is not None:
-            self.wavelengths = np.array(wavelengths)
-        else:
-            self.wavelengths = None
+    Parameters:
+    - explainer: An instance of LIMEXplainer containing the spectrum and
+      wavelength data.
+    """
+    spectrum = explainer.spectrum_instance
+    wavelengths = explainer.wavelengths
+    indices = np.arange(len(spectrum))
 
-    def set_explainer(self, explainer):
-        """
-        Sets the LIMEXplainer instance for the visualizer.
+    plt.figure(figsize=(12, 4))
+    plt.plot(indices, spectrum, label='Spectrum Instance', color='royalblue',
+             linestyle='-', linewidth=2)
+    plt.title('Spectrum Analysis', fontsize=15)
+    plt.xlabel('Wavelength (μm)', fontsize=15)
+    plt.ylabel('Scaled Values', fontsize=15)
+    plt.xticks(indices[::10], [f"{w:.2f}" for w in wavelengths[::10]],
+               rotation=45, fontsize=12)
+    plt.yticks(fontsize=12)
+    plt.grid(True, linestyle='--', linewidth=0.5, alpha=0.7)
+    plt.legend(fontsize=14)
+    plt.tight_layout()
+    plt.show()
 
-        Parameters:
-        - explainer (LIMEXplainer): An instance of LIMEXplainer class.
-        """
-        self.explainer = explainer
 
-    def plot_spectrum_instance(self, instance_id=None):
-        """
-        Plots the selected instance of a spectrum or the instance
-        specified by an ID if provided.
+def plot_perturbed_spectrum(spectrum,
+                            perturbed_spectrum,
+                            wavelengths,
+                            num_segments,
+                            perturbation):
+    """
+    Plots the spectrum divided into the specified number of segments,
+    highlighting inactive segments based on a given perturbation pattern,
+    using index-based plotting with wavelength labels.
 
-        Parameters:
-        - instance_id (int, optional): The ID of the spectrum instance to plot.
-          If not specified, uses the instance from the explainer.
-        """
-        if instance_id is not None and self.explainer:
-            spectrum = self.explainer.get_spectrum_instance_by_id(instance_id)
-        elif self.explainer:
-            spectrum = self.explainer.spectrum_instance
-        else:
-            raise ValueError(
-                "Spectrum instance not provided or set in explainer."
-            )
+    Parameters:
+    - explainer: An instance of LIMEXplainer containing all relevant
+      data and methods.
+    """
 
-        plt.figure(figsize=(12, 4))
-        if self.wavelengths is not None:
-            plt.plot(self.wavelengths, spectrum,
-                     label='Selected Spectrum Instance',
-                     color='blue', linestyle='-')
-        else:
[…]
