### 2024-05-24 #6346 @plaplant
@oparisot thank you very much for submitting your review! I'm glad to hear that you feel this is ready for acceptance.

@EhsanGharibNezhad @mwalmsley thanks for keeping up with the review! It sounds like there is just this one small issue about the package's `environment.yml` file until we're ready to move forward with acceptance. Please let me know if there is anything I can do to help move this forward.
