### 2024-06-04 9f71237
update __save_plot__ arg
