+++ environment.yml
@@ -2,255 +2,23 @@ name: TelescopeML
 channels:
   - conda-forge
 dependencies:
-  - anyio=3.7.1
-  - appnope=0.1.3
-  - argon2-cffi=21.3.0
-  - argon2-cffi-bindings=21.2.0
-  - arrow=1.2.3
   - astropy=5.3.1
-  - asttokens=2.2.1
-  - async-lru=2.0.3
-  - attrs=23.1.0
-  - babel=2.12.1
-  - backcall=0.2.0
-  - backports=1.0
-  - backports.functools_lru_cache=1.6.5
-  - beautifulsoup4=4.12.2
-  - bleach=6.0.0
   - bokeh=3.2.1
-  - brotli=1.0.9
-  - brotli-bin=1.0.9
-  - brotli-python=1.0.9
-  - bzip2=1.0.8
-  - c-ares=1.19.1
-  - ca-certificates=2023.7.22
-  - cached-property=1.5.2
-  - cached_property=1.5.2
-  - cffi=1.15.1
-  - charset-normalizer=3.2.0
-  - comm=0.1.3
-  - contourpy=1.1.0
-  - cycler=0.11.0
-  - debugpy=1.6.7
-  - decorator=5.1.1
-  - defusedxml=0.7.1
-  - entrypoints=0.4
-  - exceptiongroup=1.1.2
-  - executing=1.2.0
-  - flit-core=3.9.0
-  - fonttools=4.41.1
-  - fqdn=1.5.1
-  - freetype=2.12.1
   - h5py=3.6.0
   - hdf5=1.12.1
-  - idna=3.4
-  - importlib-metadata=6.8.0
-  - importlib-resources=6.0.0
-  - importlib_metadata=6.8.0
-  - importlib_resources=6.0.0
-  - ipykernel=6.25.0
-  - ipython=8.14.0
-  - ipython_genutils=0.2.0
-  - ipywidgets=8.0.7
-  - isoduration=20.11.0
-  - jedi=0.18.2
-  - jinja2=3.1.2
-  - joblib=1.3.0
-  - json5=0.9.14
-  - jsonschema=4.18.4
-  - jsonschema-specifications=2023.7.1
-  - jsonschema-with-format-nongpl=4.18.4
-  - jupyter=1.0.0
-  - jupyter-lsp=2.2.0
-  - jupyter_client=8.3.0
-  - jupyter_console=6.6.3
-  - jupyter_core=5.3.1
-  - jupyter_events=0.6.3
-  - jupyter_server=2.7.0
-  - jupyter_server_terminals=0.4.4
   - jupyterlab=4.0.3
-  - jupyterlab_pygments=0.2.2
-  - jupyterlab_server=2.24.0
-  - jupyterlab_widgets=3.0.8
-  - kiwisolver=1.4.4
-  - krb5=1.21.1
-  - lcms2=2.15
-  - lerc=4.0.0
-  - libblas=3.9.0
-  - libbrotlicommon=1.0.9
-  - libbrotlidec=1.0.9
-  - libbrotlienc=1.0.9
-  - libcblas=3.9.0
-  - libcurl=8.2.1
-  - libcxx=16.0.6
-  - libdeflate=1.18
-  - libedit=3.1.20191231
-  - libev=4.33
-  - libffi=3.4.2
-  - libgfortran=5.0.0
-  - libgfortran5=12.3.0
-  - libjpeg-turbo=2.1.5.1
-  - liblapack=3.9.0
-  - libnghttp2=1.52.0
-  - libopenblas=0.3.23
-  - libpng=1.6.39
-  - libprotobuf=3.19.6
-  - libsodium=1.0.18
-  - libsqlite=3.42.0
-  - libssh2=1.11.0
-  - libtiff=4.5.1
[…]
