+++ store/environment.yml
@@ -2,217 +2,215 @@ name: TelescopeML
 channels:
   - conda-forge
 dependencies:
-  - anyio=3.7.1=pyhd8ed1ab_0
-  - appnope=0.1.3=pyhd8ed1ab_0
-  - argon2-cffi=21.3.0=pyhd8ed1ab_0
-  - argon2-cffi-bindings=21.2.0=py39h0f82c59_4
-  - arrow=1.2.3=pyhd8ed1ab_0
-  - astropy=5.3.1=py39h8a366b7_0
-  - asttokens=2.2.1=pyhd8ed1ab_0
-  - async-lru=2.0.3=pyhd8ed1ab_0
-  - attrs=23.1.0=pyh71513ae_1
-  - babel=2.12.1=pyhd8ed1ab_1
-  - backcall=0.2.0=pyh9f0ad1d_0
-  - backports=1.0=pyhd8ed1ab_3
-  - backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
-  - beautifulsoup4=4.12.2=pyha770c72_0
-  - bleach=6.0.0=pyhd8ed1ab_0
-  - bokeh=3.2.1=pyhd8ed1ab_0
-  - brotli=1.0.9=h1a8c8d9_9
-  - brotli-bin=1.0.9=h1a8c8d9_9
-  - brotli-python=1.0.9=py39h23fbdae_9
-  - bzip2=1.0.8=h3422bc3_4
-  - c-ares=1.19.1=hb547adb_0
-  - ca-certificates=2023.7.22=hf0a4a13_0
-  - cached-property=1.5.2=hd8ed1ab_1
-  - cached_property=1.5.2=pyha770c72_1
-  - certifi=2023.7.22=pyhd8ed1ab_0
-  - cffi=1.15.1=py39he153c15_5
-  - charset-normalizer=3.2.0=pyhd8ed1ab_0
-  - comm=0.1.3=pyhd8ed1ab_0
-  - contourpy=1.1.0=py39hbd775c9_0
-  - cycler=0.11.0=pyhd8ed1ab_0
-  - debugpy=1.6.7=py39h23fbdae_0
-  - decorator=5.1.1=pyhd8ed1ab_0
-  - defusedxml=0.7.1=pyhd8ed1ab_0
-  - entrypoints=0.4=pyhd8ed1ab_0
-  - exceptiongroup=1.1.2=pyhd8ed1ab_0
-  - executing=1.2.0=pyhd8ed1ab_0
-  - flit-core=3.9.0=pyhd8ed1ab_0
-  - fonttools=4.41.1=py39h0f82c59_0
-  - fqdn=1.5.1=pyhd8ed1ab_0
-  - freetype=2.12.1=hadb7bae_2
-  - h5py=3.6.0=nompi_py39hd982b79_100
-  - hdf5=1.12.1=nompi_hd9dbc9e_104
-  - idna=3.4=pyhd8ed1ab_0
-  - importlib-metadata=6.8.0=pyha770c72_0
-  - importlib-resources=6.0.0=pyhd8ed1ab_1
-  - importlib_metadata=6.8.0=hd8ed1ab_0
-  - importlib_resources=6.0.0=pyhd8ed1ab_1
-  - ipykernel=6.25.0=pyh5fb750a_0
-  - ipython=8.14.0=pyhd1c38e8_0
-  - ipython_genutils=0.2.0=py_1
-  - ipywidgets=8.0.7=pyhd8ed1ab_0
-  - isoduration=20.11.0=pyhd8ed1ab_0
-  - jedi=0.18.2=pyhd8ed1ab_0
[…]
