### 2024-06-04 f9bfd5c
update TelescopeML/StatVisAnalyzer.py
