+++ store/_sources/knowledgebase.rst.txt
@@ -1,9 +1,0 @@
-KnowledgeBase
-===============
-
-Concepts: Data processing and Model Training in CNN
-----------------------------------------------------
-.. toctree::
-   :maxdepth: 1
-
-    Neural Network Arch  <tutorials/Concepts_CNN_Model.ipynb>
