+++ store/_sources/index.rst.txt
@@ -1,42 +1,0 @@
-.. TelescopeML documentation master file, created by
-   sphinx-quickstart on Tue Dec 27 15:39:09 2022.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-TelescopeML
-==============
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-``TelescopeML`` is a python library consists of a set of modules and trained ML models utilizing the developed deep Convolutional Neural Networks (CNN) and Machine Learning (ML) models to perform the following tasks:
-
-- Explore and Process the synthetic datasets (training examples)
-- Build CNN model
-- Train the module using any given hyperparameters
-
-or simply...
-
-- Load the pre-trained CNN models based on the latest synthetic datasets
-- Predict the stellar/planetaty parameters
-- Report the statistical analysis
-
-
-
-
-======================
-
-.. toctree::
-   :maxdepth: 1
-   :hidden:
-
-  
-   Installation <installation>
-   Tutorials <tutorials>
-   The Code <code>
-   KnowledgeBase <knowledgebase> 
-   Github <https://github.com/ehsangharibnezhad>
-   Publications <publications>
-   What to Cite <cite>
-
-
