+++ _sources/tutorials.rst.txt
@@ -1,40 +1,0 @@
-Tutorials
-==========
-
-
-
-Exploring the Datasets
------------------------
-.. toctree::
-   :maxdepth: 1
-
-    Brown Dwarf Synthetic Dataset  <tutorials/1__BrownDwarf_Data_Exploration.ipynb>
-
-
-
-Predict Atmospheric Parameters
--------------------------------
-.. toctree::
-   :maxdepth: 2
-
-   Predict Brown Dwarf Atmospheric Parameters  <tutorials/2__Predict_BrownDwarf_Parameters.ipynb>
-
-
-
-Data Processing
-----------------
-.. toctree::
-   :maxdepth: 2
-
-   Data Processing  <tutorials/3__Data_Processing.ipynb>
-
-
-Train a Regression CNNs Model
-------------------------------
-.. toctree::
-   :maxdepth: 2
-
-   Train a Regression CNNs Model  <tutorials/4__Train_CNN_Model.ipynb>
-
-
-
