### 2024-06-05 d5dc520
delete old sphinx docs
