+++ docs/knowledgebase.rst
@@ -1,0 +1,3 @@
+KnowledgeBase
+===============
+
