+++ docs/code.rst
@@ -1,0 +1,40 @@
+The Modules
+============
+
+
+TelescopeML\.DataMaster Module
+--------------------------------
+
+.. automodule:: TelescopeML.DataMaster
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+TelescopeML\.DeepTrainer Module
+--------------------------------
+
+.. automodule:: TelescopeML.DeepTrainer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+TelescopeML\.Predictor Module
+------------------------------------
+
+.. automodule:: TelescopeML.Predictor
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+TelescopeML\.StatVisAnalyzer Module
+------------------------------------
+
+.. automodule:: TelescopeML.StatVisAnalyzer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+
