+++ _sources/tutorials/1__BrownDwarf_Data_Exploration.ipynb.txt
@@ -98,11 +98,12 @@
     "\n",
     "We computed a low-resolution spectrum ($R$=200) utilizing atmopshric brown dwarfs grid model, [*Sonora-Bobcat*](https://arxiv.org/pdf/2107.07434.pdf) for spectral range $\\sim$0.9-2.4 $\\mu m$. An open-source atmospheric radiative transfer Python package, [*PICASO*](https://natashabatalha.github.io/picaso/) was employed for generating these datasets. This dataset encompass 30,888 synthetic spectra (or instances or rows). \n",
     "\n",
-    "Each spectrum has 104 wavelengths (i.e., 0.897, 0.906, ..., 2.512~$\\mu$m) and 4 output atmospheric parameters:\n",
-    "- gravity ( log$g$ ) \n",
-    "- temperature ( $T_{\\rm eff}$ )\n",
-    "- carbon-to-oxygen ratio ( C/O )\n",
-    "- metallicity ( [M/H] )"
+    "Each spectrum has 104 wavelengths (i.e., 0.897, 0.906, ..., 2.512 μm) and 4 output atmospheric parameters:\n",
+    "\n",
+    "- gravity (log *g*)\n",
+    "- temperature (*T*<sub>eff</sub>)\n",
+    "- carbon-to-oxygen ratio (C/O)\n",
+    "- metallicity ([M/H])\n"
    ]
   },
   {
@@ -553,10 +554,10 @@
    "source": [
     "## 1.2 Check atmospheric parameters\n",
     "\n",
-    "- gravity ( log$g$ ) \n",
-    "- temperature ( $T_{\\rm eff}$ )\n",
-    "- carbon-to-oxygen ratio ( C/O )\n",
-    "- metallicity ( [M/H] )"
+    "- gravity (log *g*)\n",
+    "- temperature (*T*<sub>eff</sub>)\n",
+    "- carbon-to-oxygen ratio (C/O)\n",
+    "- metallicity ([M/H])"
    ]
   },
   {
@@ -722,7 +723,7 @@
    "source": [
     "### 1.4.1  Effective Temperature \n",
     "\n",
-    "Plot the brown dwarf datasets for $T_{\\rm eff}$= 400-1800K while other paremeters are constnat (log$g$=5, C/O=1, [M/H]=0)"
+    "Plot the brown dwarf datasets for $T_{\\rm eff}$= 400-1800K while other paremeters are constnat (log*g*=5, C/O=1, [M/H]=0)"
    ]
   },
   {
@@ -824,7 +825,8 @@
    "source": [
     "### 1.4.3 Carbon-to-Oxygen ratio\n",
     "\n",
-    "Plot the brown dwarf datasets for C/O= 0.25-2.5 while other paremeters are constnat ($T_{eff}$=800, log$g$=5.0, [M/H]=0)"
+    "\n",
+    "Plot the brown dwarf datasets for C/O= 0.25-2.5 while other paremeters are constnat (*T*<sub>eff</sub>=800, log *g*=5.0, [M/H]=0)"
    ]
   },
   {
@@ -876,7 +878,7 @@
    "source": [
     "### 1.4.4 Metallicity\n",
     "\n",
-    "Plot the brown dwarf datasets for [M/H]= -1.0 - 2.0 while other paremeters are constnat ($T_{\\rm eff}$=800, log$g$=5.0, C/O=1.0)"
+    "Plot the brown dwarf datasets for [M/H]= -1.0 - 2.0 while other paremeters are constnat (*T*<sub>eff</sub>=800, log *g*=5.0, C/O=1.0)"
    ]
   },
   {
