+++ _sources/cite.rst.txt
@@ -1,0 +1,9 @@
+Cite:
+======
+
+(Under construction) Please cite the following papers if you are using the ``TelescopeML`` package or any of its concepts, modules, or
+functions in your work:
+
+1. Will be updated after publishing the package!
+
+2. Will be updated after publishing the first case study!
