+++ docs/_sources/knowledgebase.rst.txt
@@ -1,3 +1,0 @@
-KnowledgeBase
-===============
-
