+++ _sources/tutorials/3__Data_Processing.ipynb.txt
@@ -1,1660 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "723c394d-987d-4f2d-8bd6-378a23b1218e",
-   "metadata": {
-    "tags": []
-   },
-   "source": [
-    "# 3. Process Brown Dwarf Atmospheric Parameters \n",
-    "\n",
-    "In the following steps, you will:\n",
-    "\n",
-    "- Load the brown dwarf dataset used to train the ML models.\n",
-    "- Prepare the X and y variables to deploy the trained ML models.\n",
-    "- Visualize them for a few cases.\n",
-    "\n",
-    "We will need the following modules from `TelescopeML`:\n",
-    "\n",
-    "- **DataMaster**: to prepare the synthetic brown dwarf dataset and load the trained machine learning (ML) models.\n",
-    "- **StatVisAnalyzer**: to provide statistical tests and plotting functions.\n",
-    "- **IO_utils**: to provide functions to load the trained ML models.\n"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 1,
-   "id": "dab6ba1b-87cf-482b-9bad-7ae019708a27",
-   "metadata": {},
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "No Bottleneck unit testing available.\n"
-     ]
-    },
-    {
-     "data": {
-      "text/html": [
-       "\n",
-       "    <div class=\"bk-root\">\n",
-       "        <a href=\"https://bokeh.org\" target=\"_blank\" class=\"bk-logo bk-logo-small bk-logo-notebook\"></a>\n",
-       "        <span id=\"1001\">Loading BokehJS ...</span>\n",
-       "    </div>"
-      ]
-     },
-     "metadata": {},
-     "output_type": "display_data"
-    },
-    {
-     "data": {
-      "application/javascript": [
-       "\n",
-       "(function(root) {\n",
-       "  function now() {\n",
-       "    return new Date();\n",
-       "  }\n",
-       "\n",
-       "  var force = true;\n",
-       "\n",
-       "  if (typeof root._bokeh_onload_callbacks === \"undefined\" || force === true) {\n",
-       "    root._bokeh_onload_callbacks = [];\n",
-       "    root._bokeh_is_loading = undefined;\n",
-       "  }\n",
-       "\n",
-       "  var JS_MIME_TYPE = 'application/javascript';\n",
-       "  var HTML_MIME_TYPE = 'text/html';\n",
-       "  var EXEC_MIME_TYPE = 'application/vnd.bokehjs_exec.v0+json';\n",
-       "  var CLASS_NAME = 'output_bokeh rendered_html';\n",
-       "\n",
-       "  /**\n",
-       "   * Render data to the DOM node\n",
-       "   */\n",
-       "  function render(props, node) {\n",
-       "    var script = document.createElement(\"script\");\n",
-       "    node.appendChild(script);\n",
-       "  }\n",
-       "\n",
-       "  /**\n",
-       "   * Handle when an output is cleared or removed\n",
-       "   */\n",
-       "  function handleClearOutput(event, handle) {\n",
-       "    var cell = handle.cell;\n",
-       "\n",
-       "    var id = cell.output_area._bokeh_element_id;\n",
-       "    var server_id = cell.output_area._bokeh_server_id;\n",
-       "    // Clean up Bokeh references\n",
-       "    if (id != null && id in Bokeh.index) {\n",
-       "      Bokeh.index[id].model.document.clear();\n",
-       "      delete Bokeh.index[id];\n",
-       "    }\n",
-       "\n",
-       "    if (server_id !== undefined) {\n",
[…]
