### 2024-06-05 #6346 @EhsanGharibNezhad
Hi All,

Since the last edition, the following key updates have been made to the package and tutorials:

1. Notebooks now include flowcharts and figures, along with more detailed descriptions to provide a clear roadmap.
2. Function names in the modules are now more explicit, and some have been edited.
3. The online Sphinx documentation has been revised.
4. A new version of TelescopeML (v0.0.5) is now available on PyPI.
5. The environment file has been updated following @mwalmsley 's suggestions.

Additionally, one of my team members, Mahdi Habibi (https://github.com/mdhabibi), dedicated his time to editing the notebooks and assisting with code testing and compatibility. As a result, his name has been added to the JOSS paper draft as a co-author.

Please let me know if you have any questions. Thank you!
