### 2024-06-10 #6346 @plaplant
@mwalmsley thank you for your review! Glad to hear that you recommend acceptance.

@EhsanGharibNezhad now that both reviewers have recommended acceptance, we can move forward with the final checks before publication. In looking at [the manuscript proof from a few days ago](https://github.com/openjournals/joss-reviews/issues/6346#issuecomment-2155224971), it seems like Mahdi Habibi has not been added as an author. Please update the author list when you get a chance.

In addition to adding the author, please confirm that the latest version of TelescopeML is v0.0.5 (from your above comment). Also, please archive the software on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or some other long-term hosting platform. Once that is done, please post the DOI for the package here. Let me know if you have any questions. Thanks!
