### 2024-06-10 21ce655
update publications/joss/paper.md
