### 2024-06-10 ce73dde
update publications/joss/paper.md
