+++ publications/joss/paper.md
@@ -27,6 +27,7 @@ authors:
     affiliation: "3, 4"
   - name: Mahdi Habibi
     affiliation: "5"
+  - orcid: 0000-0001-8530-7746
   - name: Gopal Nookula 
     affiliation: "6"
 affiliations:
