### 2024-06-10 e5ae3ee
update publications/joss/paper.md
