+++ publications/joss/paper.md
@@ -26,8 +26,8 @@ authors:
     orcid: 0000-0002-2188-0807
     affiliation: "3, 4"
   - name: Mahdi Habibi
+    orcid: 0000-0001-8530-7746
     affiliation: "5"
-  - orcid: 0000-0001-8530-7746
   - name: Gopal Nookula 
     affiliation: "6"
 affiliations:
