### 2024-06-10 921af20
update publications/joss/paper.md
