### 2024-06-10 c0e3a49
update publications/joss/paper.md
