### 2024-06-10 8caebab
update publications/joss/paper.md
