### 2024-06-10 #6346 @EhsanGharibNezhad
@plaplant, I have two questions:

1. We are about to submit a scientific paper to the ApJ journal that discusses the implications of this package and the trained CNN models on brown dwarf spectra. Could you please instruct how to address that in the JOSS paper? Suggested text: "This article and software are linked with the research article DOI LINK OF ApJ PAPER - TBD, published in The Astrophysical Journal."
The title of the paper is "TelescopeML – II. Convolutional Neural Networks for Predicting Brown Dwarf Atmospheric
Parameters"

2. The footnote in the rendered PDF does not show my full name correctly. It should be "Ehsan (Sam) Gharib-Nezhad et al. (2024)....". Could you please let me know how to fix this?
