### 2024-06-12 f61fd2a
minor update to save boxplot_hist plots
