+++ TelescopeML/LIMEXplainer.py
@@ -1,0 +1,569 @@
+import os
+import sys
+import logging
+import numpy as np
+import pandas as pd
+from IPython.display import display
+from sklearn.metrics import pairwise_distances
+from sklearn.linear_model import LinearRegression
+from .StatVisAnalyzer import (plot_perturbed_spectrum,
+                              plot_most_influential_segments,
+                              plot_all_influential_segments,
+                              plot_ave_influential_segments,
+                              plot_average_influence_per_output,
+                              plot_overall_average_influence)
+
+
+class LIMEXplainer:
+    def __init__(self, model, data, features=None):
+        """
+        Initializes the LIMEXplainer with specified configurations for LIME.
+
+        Parameters:
+        - model: Trained machine learning model.
+        - data: Dataset containing input data and min-max scaled data.
+        - features (np.ndarray, optional): Array of wavelengths corresponding
+          to spectrum data points.
+        """
+        self.model = model
+        self.X_test = data
+        self.wavelengths = features
+        self.instance_index = None
+        # Default kernel width for calculating weights
+        self.kernel_width = 0.25
+        # Default perturbation function
+        self.perturb_function = self.perturb_segment_mean
+        # Default number of segments
+        self.num_segments = 10
+        # Default number of perturbations
+        self.num_perturbations = 200
+        self.spectrum_instance = None
+        self.min_max_instance = None
+        self.perturbed_spectra = None
+        self.cosine_distances = None
+        self.weights = None
+        self.surrogate_models = None
+        self.segment_importance_coefficients = None
+        self.last_perturbation_index = -1
+
+    def set_instance_index(self, instance_index):
+        """Set a new instance index and update related attributes."""
+        if not (0 <= instance_index < len(self.X_test[0][0])):
+            raise ValueError("instance_index is out of the range of the data "
+                             "provided.")
+        self.instance_index = instance_index
+        self.spectrum_instance = self.X_test[0][0][self.instance_index]
+        self.min_max_instance = self.X_test[1][0][self.instance_index]
+
+    def set_num_segments(self, num_segments):
+        """Set a new number of segments and regenerate perturbations."""
+        if num_segments is not None:
+            if num_segments < 2:
+                raise ValueError("Number of segments must be at least 2.")
+            elif num_segments > len(self.spectrum_instance):
+                raise ValueError(f"Number of segments must not exceed "
+                                 f"{len(self.spectrum_instance)}, "
+                                 f"which is the number of features "
+                                 f"in the dataset.")
+        self.num_segments = num_segments
+        self.generate_random_perturbations()
+        self.perturbed_spectra = self.perturb_and_store_spectra()
+        self.weights = self.compute_weights_from_similarity()
+        self.surrogate_models, self.segment_importance_coefficients =\
+            self.fit_surrogate_models()
+
+    def set_num_perturbations(self, num_perturbations):
+        """Set a new number of perturbations and regenerate perturbations."""
+        if num_perturbations is not None:
+            if num_perturbations < self.num_segments:
+                raise ValueError("Number of perturbations must be at least "
+                                 "equal to the number of segments.")
+            elif num_perturbations > 2 ** self.num_segments:
+                raise ValueError(f"Number of perturbations must not exceed "
+                                 f"{2 ** self.num_segments}, which is the "
+                                 f"maximum possible unique perturbations for "
+                                 f"{self.num_segments} segments.")
+        self.num_perturbations = num_perturbations
+        self.generate_random_perturbations()
+        self.perturbed_spectra = self.perturb_and_store_spectra()
+        self.weights = self.compute_weights_from_similarity()
[…]
