### 2024-06-14 fed3532
Update StatVisAnalyzer module with adding plot functions for LIME
