+++ TelescopeML/StatVisAnalyzer.py
@@ -1742,3 +1742,335 @@ def plot_regression_report(trained_ML_model,
                         bbox_inches='tight')
         plt.show()
 
+
+# ******* Data Visulaization Libraries LIME ****************************
+from matplotlib.colors import Normalize
+from matplotlib.cm import ScalarMappable
+
+
+# SpectrumVisualizer for LIMEXplainer Module
+def plot_spectrum_instance(explainer):
+    """
+    Plots the selected spectrum.
+
+    Parameters:
+    - explainer: An instance of LIMEXplainer containing the spectrum and
+      wavelength data.
+    """
+    spectrum = explainer.spectrum_instance
+    wavelengths = explainer.wavelengths
+    indices = np.arange(len(spectrum))
+
+    plt.figure(figsize=(12, 4))
+    plt.plot(indices, spectrum, label='Spectrum Instance', color='royalblue',
+             linestyle='-', linewidth=2)
+    plt.title('Spectrum Analysis', fontsize=15)
+    plt.xlabel('Wavelength (μm)', fontsize=15)
+    plt.ylabel('Scaled Values', fontsize=15)
+    plt.xticks(indices[::10], [f"{w:.2f}" for w in wavelengths[::10]],
+               rotation=45, fontsize=12)
+    plt.yticks(fontsize=12)
+    plt.grid(True, linestyle='--', linewidth=0.5, alpha=0.7)
+    plt.legend(fontsize=14)
+    plt.tight_layout()
+    plt.show()
+
+
+def plot_perturbed_spectrum(spectrum,
+                            perturbed_spectrum,
+                            wavelengths,
+                            num_segments,
+                            perturbation):
+    """
+    Plots the spectrum divided into the specified number of segments,
+    highlighting inactive segments based on a given perturbation pattern,
+    using index-based plotting with wavelength labels.
+
+    Parameters:
+    - explainer: An instance of LIMEXplainer containing all relevant
+      data and methods.
+    """
+
+    # Calculate the indices for each segment
+    indices = np.linspace(0, len(spectrum) - 1, num_segments + 1, dtype=int)
+
+    plt.figure(figsize=(12, 4))
+
+    # Plot original and perturbed spectrum using indices
+    plt.plot(spectrum, label='Spectrum Instance', color='royalblue',
+             linestyle='--', linewidth=2)
+    plt.plot(perturbed_spectrum, label='Perturbed Spectrum', color='green',
+             linestyle='-', linewidth=2)
+
+    # Highlight inactive segments
+    for i in range(num_segments):
+        if i < len(perturbation) and perturbation[i] == 0:
+            start_idx, end_idx = indices[i], indices[i + 1]
+            plt.axvspan(start_idx, end_idx, color='gray', alpha=0.5)
+
+    # Add vertical lines for segment boundaries
+    for idx in indices:
+        plt.axvline(x=idx, color='red', linestyle='--', alpha=0.75,
+                    label='Segment Boundary' if idx == indices[0] else None)
+
+    plt.title('Perturbed Spectrum with Inactive Segments Highlighted',
+              fontsize=15)
+    plt.xlabel('Wavelength (μm)', fontsize=15)
+    plt.ylabel('Scaled Values', fontsize=15)
+    # Set x-ticks to show corresponding wavelengths
+    plt.xticks(indices, [f"{wavelengths[i]:.2f}" for i in indices],
+               rotation=45, fontsize=12)
+    plt.yticks(fontsize=12)
+    plt.grid(True, linestyle='--', linewidth=0.5, alpha=0.7)
+    plt.legend(fontsize=14)
+    plt.tight_layout()
+    plt.show()
+
+
+def plot_most_influential_segments(spectrum, wavelengths,
+                                   number_of_top_features,
[…]
