### 2024-06-18 a8a2078
Make some small editorial changes
