+++ publications/joss/paper.bib
@@ -508,17 +508,6 @@ archivePrefix = {arXiv},
 
 
 
-@article{batalha2017pandexo,
-  title={PandExo: a community tool for transiting exoplanet science with JWST \& HST},
-  author={Batalha, Natasha E and Mandell, Avi and Pontoppidan, Klaus and Stevenson, Kevin B and Lewis, Nikole K and Kalirai, Jason and Earl, Nick and Greene, Thomas and Albert, Lo{\"\i}c and Nielsen, Louise D},
-  journal={Publications of the Astronomical Society of the Pacific},
-  volume={129},
-  number={976},
-  pages={064501},
-  year={2017},
-  publisher={IOP Publishing}
-}
-
 @article{batalha2019picaso,
   title={Exoplanet reflected-light spectroscopy with PICASO},
   author={Batalha, Natasha E and Marley, Mark S and Lewis, Nikole K and Fortney, Jonathan J},
@@ -530,25 +519,6 @@ archivePrefix = {arXiv},
   doi = {10.3847/1538-4357/ab1b51},
   publisher={IOP Publishing}
 }
-
-@article{bean2018transiting,
-  title={The transiting exoplanet community early release science program for JWST},
-  author={Bean, Jacob L and Stevenson, Kevin B and Batalha, Natalie M and Berta-Thompson, Zachory and Kreidberg, Laura and Crouzet, Nicolas and Benneke, Bj{\"o}rn and Line, Michael R and Sing, David K and Wakeford, Hannah R and others},
-  journal={Publications of the Astronomical Society of the Pacific},
-  volume={130},
-  number={993},
-  pages={114402},
-  year={2018},
-  publisher={IOP Publishing}
-}
-
-@article{MacDonald2023POSEIDON,
-doi = {10.21105/joss.04873},
-url = {https://doi.org/10.21105/joss.04873},
-year = {2023}, publisher = {The Open Journal},
-volume = {8}, number = {81}, pages = {4873}, author = {Ryan J. MacDonald},
-title = {POSEIDON: A Multidimensional Atmospheric Retrieval Code for Exoplanet Spectra},
-journal = {Journal of Open Source Software} }
 
 @ARTICLE{Valizadegan2021,
        author = {{Valizadegan}, Hamed and {Martinho}, Miguel and {Wilkens}, Laurent S. and {Jenkins}, Jon M. and {Smith}, Jeffrey and {Caldwell}, Douglas A. and {Twicken}, Joseph D. and {Gerum}, Pedro C. and {Walia}, Nikash and {Hausknecht}, Kaylie and {Lubin}, Noa Y. and {Bryson}, Stephen T. and {Oza}, Nikunj C.},
