### 2024-06-25 #6346 @EhsanGharibNezhad
@plaplant thanks for responding to my questions.

I think we do not need to wait for the ApJ paper and it is okay to go ahead and publish this one. I'll do a final check this week and will let you know if there is any typo or change. Thanks!
