### 2024-06-25 50f34de
update joss - few minor edits
