### 2024-06-25 1a55c47
update joss - few minor edits
