### 2024-06-25 6d5aafb
update joss - few minor edits
