### 2024-06-25 ec68875
update joss - update the info figure
