### 2024-06-25 #6346 @EhsanGharibNezhad
> 10.5281/zenodo.11553655

> @EhsanGharibNezhad thanks for your patience and your work on getting this over the finish line! I made a few suggestions for changes to the paper in a PR ([EhsanGharibNezhad/TelescopeML#230](https://github.com/EhsanGharibNezhad/TelescopeML/pull/230)). To answer your questions above:
>
> > 1. We are about to submit a scientific paper to the ApJ journal that discusses the implications of this package and the trained CNN models on brown dwarf spectra. Could you please instruct how to address that in the JOSS paper? Suggested text: "This article and software are linked with the research article DOI LINK OF ApJ PAPER - TBD, published in The Astrophysical Journal."
> >    The title of the paper is "TelescopeML – II. Convolutional Neural Networks for Predicting Brown Dwarf Atmospheric
> >    Parameters"
>
> JOSS does have a partnership with AAS publishing for companion papers, which is outlined more [here](https://joss.readthedocs.io/en/latest/editing.html#aas-publishing). My reading of this partnership is that the joint publication should have been established earlier in the review process, before the review of the JOSS paper began so that the reviewers could be made aware. The metadata tags `aas-doi` and `aas-journal` are used for this type of joint publication, so I think they do not apply in this case and should probably be removed from the paper.
>
> Would it be possible to submit your article to a pre-print server (e.g., arXiv) and link the DOI to that? If you would prefer to wait for the DOI of the published ApJ article, we would most likely have to pause the publication of the JOSS paper until the ApJ article has been accepted. Please let me know how you'd like to proceed.

- At this stage, I'd prefer to wait until I get the reviewers' revision and then post it in the arXiv. I added this statement in the summary but please feel free to take it out if you think it is out of place:

**The implications and scientific outcomes from the trained CNN models and this package are under revision for The
Astrophysical Journal under the title *TelescopeML – II: Convolutional Neural Networks for Predicting Brown Dwarf
Atmospheric Parameters*.**

> > 2. The footnote in the rendered PDF does not show my full name correctly. It should be "Ehsan (Sam) Gharib-Nezhad et al. (2024)....". Could you please let me know how to fix this?
>
> The PR I opened on the paper has a suggestion that should fix this. Thanks for pointing it out!
>
> Let me know if you have any questions!

Fixed! Thanks!
