### 2024-06-25 #6346 @EhsanGharibNezhad
@plaplant

Please find the new version of the JOSS draft with the following changes:

- All comments in the PR are applied
- The author name is correct now in the footnote and the author list
- Few grammatical errors were fixed.
