### 2024-07-04 #6346 @plaplant
@EhsanGharibNezhad thanks for addressing the remaining issues! I made a small PR [here](https://github.com/EhsanGharibNezhad/TelescopeML/pull/231) that should resolve a few small citation and reference issues. Once that's merged in, I think this will be ready to accept!
