### 2024-07-08 #6346 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
