### 2024-07-10 #6346 @EhsanGharibNezhad
> 👋 @openjournals/aass-eics, this paper is ready to be accepted and published.
>
> Check final proof 👉📄 [Download article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06346/joss.06346/10.21105.joss.06346.pdf)
>
> If the paper PDF and the deposit XML files look good in [openjournals/joss-papers#5587](https://github.com/openjournals/joss-papers/pull/5587), then you can now move forward with accepting the submission by compiling again with the command `@editorialbot accept`

@plaplant I'm not sure if I need to compile that command?
