### 2024-07-18 #6346 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
