### 2024-07-25 #6346 @EhsanGharibNezhad
> Many thanks to @oparisot and @mwalmsley for reviewing and to @plaplant for editing! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!
>
> @EhsanGharibNezhad — Your paper is now accepted and published in JOSS! ⚡🚀💥
>
> Note: I'm currently seeing that the PDF isn't appearing on the DOI page. This sometimes happens for a few hours because of some caching issues on the back end. If this doesn't resolve itself by tomorrow, let me know and I'll look into it!

Thank you all! It was a wonderful experience going through all steps! :)
